% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.rate.R
\name{adjust.multiplier}
\alias{adjust.multiplier}
\title{Convert between multipliers of the same unit, e.g. mg to kg}
\usage{
adjust.multiplier(x, input, output)
}
\arguments{
\item{x}{numeric.}

\item{input}{string.}

\item{output}{string.}
}
\value{
A numeric.
}
\description{
This is an internal function. Converts units of the same scale, e.g. mg to
kg, or mL to L.
}
\keyword{internal}
