% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{adjust_rate}
\alias{adjust_rate}
\title{Simple rate adjustment function for background correction or open respirometry oxygen flux}
\usage{
adjust_rate(df, by)
}
\arguments{
<<<<<<< HEAD
\item{df}{data frame, or any object of class \code{calc_rate} or \code{auto_rate}}

\item{by}{either a numeric, or an object of class \code{calc_rate.bg}.}
=======
\item{df}{data frame, or any object of class \code{calc_rate} or \code{auto_rate}. This object should contain the rate that has not been adjusted, yet.}

\item{by}{either a numeric, or an object of class \code{calc_rate.bg}. This is the value to adjust the value with.}
>>>>>>> develop
}
\value{
A list object of class \code{adjust_rate}.
}
\description{
A simple rate correction function. \strong{Important:} Please note that the sign of the number must be considered in the correction. Respiration is normally a negative number, while oxygen flux may or may not be negative!
}
<<<<<<< HEAD
\examples{
adjust_rate(7.44, -0.04) # this is simply 7.44 - 0.04 = 7.40

bg <- calc_rate.bg(urchins.rd, xcol = 1, ycol = 18:19)
rt <- calc_rate(inspect_data(urchins.rd, 1, 5))
adjust_rate(rt, bg)
=======
\details{
Users may want to perform rate corrections due to background respiration or to account for oxygen flux in open respirometry experiments. We have provided a simple function to perform either, or both, corrections. This function is open to improvements and enhancements (thus, it is a separate function even though it performs a relatively rudimentary calculation).
}
\examples{
adjust_rate(7.44, -0.04) # this is simply 7.44 - 0.04 = 7.40
>>>>>>> develop
}
