% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_funs.R
\name{adjust_scale}
\alias{adjust_scale}
\title{Convert between multipliers of the same unit, e.g. mg to kg}
\usage{
adjust_scale(x, input, output)
}
\arguments{
\item{x}{numeric.}

\item{input}{string.}

\item{output}{string.}
}
\value{
A numeric.
}
\description{
Converts units of the same scale, e.g. mg to kg, or mL to L.
}
\keyword{internal}
