% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{algae.rd}
\alias{algae.rd}
\title{Oxygen production respirometry data}
\format{
A data frame object consisting of 1200 rows (20 h of data),and
2 columns: \verb{$Time} in hours, \verb{$Oxygen} in \% air saturation.
}
\usage{
algae.rd
}
\description{
Data from a respirometry experiment on algae which shows oxygen production
over time.
}
\details{
\itemize{
\item Dissolved oxygen units: \verb{\% Air Saturation}
\item Time units: \code{hours}
\item Respirometer volume (L): \code{0.1}
\item Temperature (°C): \code{12}
\item Salinity: \code{30}
}
}
\author{
Nicholas Carey
}
\keyword{datasets}
