% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.int.R
\name{auto_rate.rep}
\alias{auto_rate.rep}
\title{Function for auto_rate with replicate number in summary plus limit to n
results}
\usage{
auto_rate.rep(
  x,
  method = "linear",
  width = NULL,
  by = "row",
  plot = TRUE,
  rep = 1,
  n = 1,
  rep_row = NULL,
  meas_row = NULL,
  meas_endrow = NULL,
  rep_data = NULL,
  ...
)
}
\arguments{
\item{x}{df or inspect obj}

\item{method}{auto_rate method}

\item{width}{auto_rate width}

\item{by}{auto_rate by}

\item{plot}{auto_rate plot}

\item{rep}{The replicate}

\item{n}{number of results to return}

\item{rep_row}{start row of this rep in the larger dataset}

\item{meas_row}{start row of the measure phase in \emph{this} rep}

\item{rep_data}{df of the entire replicate. Used in plotting.}

\item{...}{pass plot stuff}
}
\value{
a \code{auto_rate} object with \code{summary$rep} filled in with appropriate
replicate number
}
\description{
This is an internal function for \code{auto_rate.int()}. Runs \code{auto_rate} and
replaces the \verb{$rep} column in \verb{$summary} with the \code{rep} input. Also limits
number of results returned using \code{n}.
}
\keyword{internal}
