% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{auto_rate_interval}
\alias{auto_rate_interval}
\title{Obtain rate values at non-overlapping intervals of a dataset}
\usage{
auto_rate_interval(dt, width, by)
}
\arguments{
\item{dt}{data.frame object.}

\item{width}{numeric.}

\item{by}{string.}
}
\value{
a list object with appended class \code{auto_rate_interval}
}
\description{
This is an internal function for \code{auto_rate()}
}
\keyword{internal}
