% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{auto_rate_linear}
\alias{auto_rate_linear}
\title{Linear detection method}
\usage{
auto_rate_linear(dt, width, by, verify = TRUE)
}
\arguments{
\item{dt}{data.frame object.}

\item{width}{numeric.}

\item{verify}{logical. Should KDE be performed again to verify the detection?}
}
\value{
a list object
}
\description{
This is an internal function for \code{auto_rate()}
}
\keyword{internal}
