% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{auto_rate_max}
\alias{auto_rate_max}
\title{Perform rolling regression and rank from NUMERICAL maximum to minimum
i.e. includes sign - most positive rates are highest}
\usage{
auto_rate_max(dt, width, by = "row")
}
\arguments{
\item{dt}{data.frame object.}

\item{width}{numeric.}

\item{by}{string.}
}
\value{
a list object with appended class \code{auto_rate_max}
}
\description{
This is an internal function for \code{auto_rate()}
}
\examples{
NULL
}
\keyword{internal}
