% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{auto_rate_min}
\alias{auto_rate_min}
\title{Perform rolling regression and rank from maximum to minimum}
\usage{
auto_rate_min(dt, width, by = "row")
}
\arguments{
\item{dt}{data.frame object.}

\item{width}{numeric.}

\item{by}{string.}
}
\value{
a list object with appended class \code{auto_rate_min}
}
\description{
This is an internal function for \code{auto_rate()}
}
\examples{
NULL
}
\keyword{internal}
