% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background.rd}
\alias{background.rd}
\title{Background respirometry data}
\format{A data frame object consisting of 20664 rows (approx 6 h of
data),and 2 columns: time in seconds, oxygen in \% air saturation.}
\usage{
background.rd
}
\description{
Background respirometry recording. Data is from a Loligo swim tunnel, oxygen
recorded via a Witrox sensor in \% air saturation over nearly 6 hours at 1
second intervals.
}
\details{
\itemize{
\item Dissolved oxygen units: \% Air Saturation
\item Time units: seconds
\item Swim tunnel volume: 12.3 L
\item Temperature: 14.5 C
\item Salinity - 34 ppt
}
}
\author{
Nicholas Carey
}
\keyword{datasets}
