% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background_exp.rd}
\alias{background_exp.rd}
\title{Background respirometry data (exponential)}
\format{
A data frame object consisting of 20664 rows (approx 6 h of data),and
2 columns: time in seconds, oxygen in \% air saturation.
}
\usage{
background_exp.rd
}
\description{
Exponential background oxygen consumption data. Data shows a generally
increasing background rate. Taken from a Loligo swim tunnel background
recording. Oxygen recorded via a Witrox sensor in \% air saturation over
nearly 6 hours at 1 second intervals. Data is from a real experiment, but
oxygen decrease curve has been exaggerated to impose an exponential increase
in background consumption for testing purposes.
}
\details{
\itemize{
\item Dissolved oxygen units: \% Air Saturation
\item Time units: seconds
\item Swim tunnel volume: 12.3 L
\item Temperature: 14.5 C
\item Salinity: 34 ppt
}
}
\author{
Nicholas Carey
}
\keyword{datasets}
