% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_funs.R
\name{broken_stick}
\alias{broken_stick}
\title{Perform broken-stick regressions}
\usage{
broken_stick(dt, n)
}
\value{
a data.table object
}
\description{
Perform broken-stick regressions
}
\keyword{internal}
