% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.mo2.R
\name{calc.mo2}
\alias{calc.mo2}
\title{Calculate volume and/or mass-specific rate of oxygen flux}
\usage{
calc.mo2(x, unit.in, unit.out, volume = NULL, mass = NULL, rank = 1,
  S = 35, t = 25, P = 1.013253)
}
\arguments{
\item{x}{Number, or an object of class \code{'calc.rate'}}

\item{unit.in}{Character.}

\item{unit.out}{Character.}

\item{volume}{Numeric. Volume of water in L.}

\item{mass}{Numeric. Mass/weight of specimen in kg.}

\item{S}{Numeric. Salinity in PSU. Defaults to 35.}

\item{t}{Numeric. temperature in degrees C. Defaults to 25.}

\item{P}{Numeric. Pressure. Defaults to 1.013253.}
}
\value{
A conversion from raw rate to volume and/or mass-specific rate.
}
\description{
Calculate volume and/or mass-specific rate of oxygen flux
}
\examples{
TBC
}
