% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.ft.R
\name{calc_rate.ft}
\alias{calc_rate.ft}
\title{Calculate flowthrough rate of change in dissolved oxygen}
\usage{
calc_rate.ft(x = NULL, time = NULL, inflow.o2 = NULL, outflow.o2 = NULL,
  flowrate = NULL, plot = TRUE)
}
\arguments{
\item{x}{data frame or object of class \code{inspect_data}. Defaults to NULL. Will
process a data frame if it is provided here.}

\item{time}{numeric. Defaults to NULL. This selects the time column if
a data frame ('df') is provided. Otherwise, this is a numeric vector for
time data.}

\item{inflow.o2}{numeric. Defaults to NULL. This selects the incurrent column if
a data frame ('df') is provided. Otherwise, this is a numeric vector for
incurrent oxygen concentration.}

\item{outflow.o2}{numeric. Defaults to NULL. This selects the outcurrent column
if a data frame ('df') is provided.  Otherwise, this is a numeric vector
for extcurrent oxygen concentration.}

\item{flowrate}{numeric vector. The flow rate. No unit of measurement is
expected; you will specify it when you perform conversions later on.}

\item{plot}{logical. Defaults to TRUE. Plots the data.}
}
\value{
An object of class "calc_rate.ft".
}
\description{
Calculates rate of $O_2$ uptake in flowthrough respirometry given a flow-rate
and both inflow.o2 and outflow.o2 oxygen concentrations.
}
\details{
Can return a single value, or multiple and mean values based on continuous
data.
}
\examples{
NULL
}
