% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.int.R
\name{calc_rate.rep}
\alias{calc_rate.rep}
\title{Function for calc_rate with replicate number in summary}
\usage{
calc_rate.rep(
  x,
  from = NULL,
  to = NULL,
  by = "time",
  plot = TRUE,
  rep = 1,
  supp.mess = TRUE,
  ...
)
}
\arguments{
\item{x}{df or inspect obj}

\item{from}{calc_rate from}

\item{to}{calc_rate to}

\item{by}{calc_rate by}

\item{plot}{calc_rate plot}

\item{rep}{The replicate}

\item{supp.mess}{Suppress messages?}
}
\value{
a \code{calc_rate} object with \code{summary$rep} filled in with appropriate
replicate number
}
\description{
This is an internal function for \code{calc_rate.int()}. Runs \code{calc_rate}
and replaces the \verb{$rep} column in \verb{$summary} with the \code{rep}
input.
}
\keyword{internal}
