% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{calc_rolling_win}
\alias{calc_rolling_win}
\title{Automatically calculate rolling window}
\usage{
calc_rolling_win(dt, width, by)
}
\arguments{
\item{dt}{data.frame object.}

\item{width}{numeric.}

\item{by}{string.}
}
\value{
a numeric
}
\description{
The calculated value is used to determine the rolling window for rolling
regressions.
}
\examples{
NULL
}
\keyword{internal}
