% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{calc_win}
\alias{calc_win}
\title{Automatically calculate rolling window}
\usage{
calc_win(dt, width, by, msg)
}
\arguments{
\item{dt}{data.frame object.}

\item{width}{numeric.}

\item{by}{string.}

\item{msg}{string. Attach function name for message (usually auto_rate)}
}
\value{
a numeric
}
\description{
The calculated value is used to determine the rolling window for rolling
regressions.
}
\keyword{internal}
