% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{calc_window}
\alias{calc_window}
\title{Calculate rolling window size for rolling regression}
\usage{
calc_window(dt, width, by)
}
\description{
Calculate rolling window size for rolling regression
}
\keyword{internal}
