% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMO2.R
\name{calcMO2}
\alias{calcMO2}
\title{Calcurate MO2}
\usage{
calcMO2(input, from, to, volume, mass = NULL, ...)
}
\arguments{
\item{input}{A number, or an object of class \code{calcRate}. You may directly parse the output of a previous \code{calcRate} result.}

\item{from}{Character string. This are the units used to calculate the rate in \code{input}. Must include the precise O2 unit and the time unit, e.g. "mg L-1 s-1" or "mg/l/s". A space, "/" or "." must separate each unit in the character string.}

\item{to}{Character string. Your intended output units. Mass, if included, must be positioned at the end of the string e.g. "mg/l/kg", NOT "mg/kg/l". A space, "/" or "." must separate each unit in the character string.}

\item{volume}{Numeric. Volume of the seawater used, in litre/liter (L).}

\item{mass}{Numeric. Defaults to NULL. Mass/weight of the animal, in kilograms (kg).}

\item{...}{for s, P, t. I'll probably convert this soon (just testing the ... method but it gets unclear)}
}
\value{
I'll put something in here soon.
}
\description{
Still working out a good description for this...
}
\examples{
# Basic conversion, with no specimen weight:
calcMO2(0.5, 'mg l-1s-1', 'mg/s', volume = 1)

# Below, will output an error as the "to" argument is wrongly formatted.
# Instead of 'ug/kg/h", the mass should be last, i.e. 'ug/h/kg'
calcMO2(0.0007, from = '\%/s', to = 'ug/kg/h', volume = 0.74, mass = 0.33)
calcMO2(0.0007, '\%/s', 'ug/h/kg', volume = 0.74, mass = 0.33) # this works

# We can use this function directly on an output of "calcRate"
data("sardine")
sard <- calcRate(sardine)
sard_mo2 <- calcMO2(sard, '\% s-1', 'ug/h/kg', volume = 1.1, mass = 0.21)

# Intermittent data is also supported
data("intermittent")
intermm <- calcRate(intermittent, c(200,2300,4100), c(1800,3200,4600), by = 'time')
intermm_mo2 <- calcMO2(intermm, from = 'mg/l/s', to = 'mg/s/kg', volume = 0.8, mass = 0.3)
}
