% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRate.R
\name{calcRate}
\alias{calcRate}
\title{Calculate rate via local linear regression or 2-point slope.}
\usage{
calcRate(df, from = NULL, to = NULL, by = "time")
}
\arguments{
\item{df}{Dataframe.}

\item{from}{Number, or a list.}

\item{to}{Number, or a list.}

\item{by}{Character. Select method of subsetting. If by \code{'time'}(default) or \code{'row'}, \code{calcRate} selects the values inclusive. If by \code{'o2'}, the first occurrence of the O2 value in \code{'from'}, and the last occurrence of the O2 value in \code{'to'}, are selected. If by \code{'proportion'}, he proportion, based on the minimum and maximum oxygen concentration. At 1, this selects the highest known oxygen value. At 0, this selects the lowest known oxygen value. Note that due to the noisy nature of respiration data, the lowest known oxygen value may not represent the last row of the dataframe.}
}
\value{
A summary list.
}
\description{
\code{calcRate} parses a dataframe, or its subset(s) to calculate the rate of change of oxygen concentration over time.
}
\examples{
data(sardine)
calcRate(sardine, 200, 1800)                 # default subset by 'time'
calcRate(sardine, 93, 92, by = 'o2')         # subset by O2
calcRate(sardine, 200, 1800, by = 'row')     # subset by row
calcRate(sardine, .8, .2, by = 'proportion') # subset by proportion

# Using a list in 'from' and 'to' calculates multiple regressions:
data(intermittent)
calcRate(intermittent, c(200,2300,4100), c(1800,3200,4600), by = 'time')

}
