% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.do.R
\name{checkUnits}
\alias{checkUnits}
\title{Measurement unit check and identification}
\usage{
checkUnits(unit, type)
}
\arguments{
\item{unit}{Character string. A unit of measurement.}

\item{type}{Numeric. Identifies the type of unit to check. 1: time; 2: O~2~
3: volume; 4: mass.}
}
\description{
Checks input string against a known database and returns a
logical output (i.e. TRUE or FALSE). If TRUE, also prints out the identified
string. Works for time (e.g min, sec), unit of O2 measurement (e.g. %, mg/L),
volume (e.g. mL, L) and weight/mass (e.g. kg).
}
\details{
This function makes it easier to parse and understand the plethora
of units that a user may provide as inputs while using our functions. The
user may deploy a 'fuzzy' approach during the input of measurement units
(e.g. 'mg/l', 'mgL-1', 'mgl-1' and 'mg per litre' all work).
}
\examples{
checkUnits('min', 'time')
checkUnits('notatunit', 'time')

checkUnits('hpa', 'o2')
checkUnits('notatunit', 'o2')
}
\author{
Januar Harianto
}
