% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.do.R
\name{convert.do}
\alias{convert.do}
\title{Convert units of rate of change in oxygen concentration over time}
\usage{
convert.do(x, from, to, S = 35, t = 25, P = 1.013253)
}
\arguments{
\item{x}{numeric vector. The number to convert.}

\item{from}{string. The unit before conversion.}

\item{to}{string. The unit after conversion.}

\item{S}{numeric. Salinity, defaults to 35.}

\item{t}{numeric. Temperature in degrees C, defaults to 25.}

\item{P}{numeric. Pressure unit in bar, defaults to 1.013253.}
}
\value{
a numeric vector.
}
\description{
Basic conversion function for non- volume- or weight-specific rate of change
in oxygen concentration over time.
}
\details{
The user may use \link{calc.mo2} for conversions involving volume- or
weight-specific rate of change in oxygen concentration over time.
}
\examples{
convert.do(90, from = '\%', to = 'mg/l')
convert.do(7.7, from = 'mgl-1', to = 'umol/l', S = 30, t = 20)

# convert a vector, if necessary
convert.do(sardine[2], from = '\%', to = 'mg/l')

}
\author{
Januar Harianto & Nicholas Carey
}
