% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.rate.R
\name{convert.rate}
\alias{convert.rate}
\title{Convert units of dissolved oxygen.}
\usage{
convert.rate(x, from = NULL, to = NULL, S = 35, t = 25, P = 1.013253)
}
\arguments{
\item{x}{numeric vector or object of class \code{calc.rate}, \code{auto.rate} or \code{adjust.rate}.}

\item{from}{string.}

\item{to}{string.}

\item{S}{numeric. Salinity. Defaults to 35.}

\item{t}{numeric. Temperature. Defaults to 25 (degrees celcius).}

\item{P}{numeric. Pressure. Defaults to 1.013253 (kPa).}
}
\value{
A list.
}
\description{
Convert units of dissolved oxygen.
}
\examples{
# Convert an object of class `calc.rate`:
x <- calc.rate(sardine.rd)
convert.rate(x, from = "\%", to = "mg/L")

# Convert a numeric:
convert.rate(100, from = "\%", to = "mg/L", S = 33, t = 18)

Convert a numeric vector:
convert.rate(sardine.rd[[2]], from = "\%", to = "torr")
}
