% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time.R
\name{convert_time}
\alias{convert_time}
\title{Parse datetime data to numeric}
\usage{
convert_time(x, format = "ymd_HMS", start = 0)
}
\arguments{
\item{x}{vector or dataframe object. Your input data. Date-time data should
be in the first column, and can be in any POSIXct or character string
format.}

\item{format}{a character string of date-time formats recognised by
\code{lubridate}. Extended formats are also supported, based on
\code{parse_date_time()}.}

\item{start}{numeric. Defaults to \code{0}. At what time (in seconds) should the
formatted time data start?}
}
\value{
A data frame object.
}
\description{
\code{convert_time()} is a function to parse class POSIX.ct or text strings of
date-time data to numeric time for use in \code{respR} functions.
}
\details{
With the help of the excellent \pkg{lubridate} package, you can simplify time
conversion of dates and time into numeric format by speficying the formats in
the \code{format} argument, without the need to include separators and % prefix.

\describe{
\item{\code{d}}{Day of the month as decimal number (01--31 or 0--31).}
\item{\code{m}}{Month of the year as decimal number (01--12 or 1--12).}
\item{\code{y}}{Year (2010, 2001, 1989).}
\item{\code{H}}{Hour, must be capitalised (\code{H}, not \code{h}). Decimal number (00--24 or 0--24).}
\item{\code{M}}{Minute, must be capitalised (\code{M}, not \code{m}). Decimal number (00--59 or 0--59).}
\item{\code{S}}{Second, must be capitalised (\code{S}, not \code{s}). Decimal number (00--59 or 0--59).}
\item{\code{p}}{AM/PM indicator. Adding this will format the data as 12-h date-time format.}
}
}
\examples{

# convert year-month-day hour-min-sec
x <- c("09-02-03 01:11:11", "09-02-03 02:11:11","09-02-03 02:25:11")
convert_time(x)
## [1]    0 3600 4440

# convert day-month-year hour-min
x <- c("03-02-09 01:11", "03-02-09 02:11","03-02-09 02:25")
convert_time(x, format = "dmyHM")
convert_time(x, format = "dmy HM")
convert_time(x, format = "dmy_HM")
convert_time(x, format = "d m y H M")
## [1]    0 3600 4440

# convert when AM/PM is present
x <- c("09-02-03 11:11:11 AM", "09-02-03 12:11:11 PM","09-02-03 01:25:11 PM")
convert_time(x, format = "dmyHMS") # this is wrong
convert_time(x, format = "dmyHMSp")
convert_time(x, format = "dmy HMS p")
convert_time(x, format = "dmy_HMS_p")
convert_time(x, format = "d m y H M S p")
## [1]    0 3600 8040

# convert dataframe with year-month-day hour-min-sec
x <- data.frame(
  x = c("09-02-03 01:11:11", "09-02-03 02:11:11","09-02-03 02:25:11"),
  y = c(23, 34, 45))
convert_time(x)

# convert dataframe with multiple columns and non-default format
x <- data.frame(
  x = c("09-02-03 11:11:11 AM", "09-02-03 12:11:11 PM","09-02-03 01:25:11 PM"),
  y = c(23, 34, 45),
  z = c(56, 67, 78))
convert_time(x, format = "dmyHMSp")

}
