% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convUnits.R
\name{convUnits}
\alias{convUnits}
\title{Convert units or oxygen concentration or respiration}
\usage{
convUnits(x, from, to, S = 35, t = 25, P = 1.013253)
}
\arguments{
\item{x}{Numeric vector of measurement(s).}

\item{from}{String. The unit to convert from.}

\item{to}{String. The unit to convert into.}

\item{S}{Numeric. Salinity, default at S = 35.}

\item{t}{Numeric. Temperature, default at 25 &deg;C.}

\item{P}{Numeric. Pressure. Default at 1.013253 bar.}
}
\description{
This function is able to pick up whether the data that needs to
be converted is a unit of [O2] or a unit of VO2 or MO2.
}
\details{
While you could parse an entire raw dataset with this, it is highly
recommended that unit conversion is conducted at the end of analysis (i.e.
convert one value once, rather than convert an entire data frame at start),
}
\examples{
convUnits(df, from = '\%', to = 'mg/l')
convUnits(df, from = 'mgl-1', to = 'umol/l', S = 30, t = 20)
}
