% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corrupted1k}
\alias{corrupted1k}
\title{Output object for the function \code{test_lin()} for "corrupted" data}
\format{A list object containing a dataframe object and an \code{lm} object.}
\usage{
corrupted1k
}
\description{
This is the output data for the performance check on \code{auto_rate()}'s linear
detection algorithm. Data was randomly-generated, then tested, using the code
\code{test_lin(reps = 1000, len = 300, type = "corrupted")}. Contain 1,000
iterations of the same test. For more information about the properties of
"corrupted" data, please read our vignette on \code{auto_rate()}.
}
\author{
Januar Harianto
}
\keyword{datasets}
