% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{drollplot}
\alias{drollplot}
\title{Generate density of rolling regression plot}
\usage{
drollplot(obj, pos = 1)
}
\arguments{
\item{obj}{object of class \code{auto_rate}.}

\item{pos}{numeric. Ranked result.}
}
\description{
Generate density of rolling regression plot
}
\keyword{internal}
