% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funs.R
\name{errorCheck}
\alias{errorCheck}
\title{Wrapper for if-stop condition in functions.}
\usage{
errorCheck(condition, msg)
}
\arguments{
\item{condition}{The expression/condition to evaluate.}

\item{msg}{An error message which will be printed if condition is fulfiled.}
}
\description{
A very simple function that may not be used...
}
\examples{
errorCheck(x > 5, "Value is bigger than 5.")
}
\author{
Januar Harianto
}
