% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.R
\name{extract_indices}
\alias{extract_indices}
\title{Extract row, time and DO indices from a subset dataframe}
\usage{
extract_indices(x, subsets, n)
}
\arguments{
\item{x}{data frame.}

\item{subsets}{list of data frames.}

\item{n}{numeric. Choose which subset in the list to extract data from.}
}
\value{
A `data.table`` object.
}
\description{
This is an internal function. Extracts row, time and DO values from a data
subset in a list.
}
\keyword{internal}
