% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rate.R
\name{filter_rate}
\alias{filter_rate}
\title{Filter auto_rate results based on a range of criteria}
\usage{
filter_rate(x, method = NULL, n = 1, plot = TRUE)
}
\arguments{
\item{x}{list. An object of class \code{auto_rate} or \code{auto_rate_filt}.}

\item{method}{string. Method by which to filter rate results. Matching results
are \emph{retained} in the output. See Details.}

\item{n}{numeric. Number, percentile, or range of results to return. See
Details.}

\item{plot}{logical. Plots a summary of all filtered plot locations within
data.}
}
\value{
A list object of class \code{auto_rate} and \code{auto_rate_filt}.
}
\description{
The \code{auto_rate} function is a very powerful function, but the
output can be large and difficult to explore. \code{filter_rate} filters
\code{auto_rate} results according to various criteria. For example, extracting
only positive or negative rates, or only the highest or lowest rates.

Note: when choosing a \code{method}, keep in mind that to keep mathematically
consistent, \code{respR} outputs oxygen consumption (i.e. respiration) rates as
negative values. This is particularly important in the difference between
\code{highest/lowest} and \code{min/max} methods.

When a \verb{$rate} is omitted by the filtering criteria, all associated data in
\verb{$summary} (i.e. the associated row) is removed. Summary results are not
reordered in any way, only the non-matching rates removed. The \verb{$peaks} and
\verb{$metadata$no_peaks} components of the \code{auto_rate} object are similarly
adjusted. The resulting object retains the \code{auto_rate} class, however an
additional \code{auto_rate_filt} class is applied.

Multiple filtering criteria can be applied by chaining (via \verb{\%>\%} piping)
the function multiple times, or assigning and running the output through it
multiple times with the different criteria. See examples.
}
\details{
These are the current methods by which \code{auto_rate} objects can be
filtered:

\subsection{ }{\code{positive}, \code{negative}} Retains all \code{positive} (>0) or
\code{negative} (<0) rates. \code{n} is ignored. Useful in intermittent respirometry
where \code{auto_rate} may identify linear regions of oxygen increase during
flushes. Note, to keep mathematically consistent, \code{respR} outputs oxygen
consumption (i.e. respiration) rates as negative values.

\subsection{ }{\code{nonzero}, \code{zero}} Retains all \code{nonzero} rates (i.e. removes
any zero rates), or retains \emph{only} \code{zero} rates (i.e. removes all rates with
any value). \code{n} is ignored.

\subsection{ }{\code{lowest}, \code{highest}} Special note should be taken of these
methods. They can only be used when rates all have the same sign (i.e. are
all negative or all positive). Here, they retain the highest and lowest
\emph{\strong{absolute}} rate values. That is, if rates are all negative, \code{method = 'highest'} will retain the highest magnitude rates regardless of sign, that
is the \emph{most negative}. Essentially, these options ignore the sign of the
rate. \code{n} indicates how many of the lowest/highest rates to extract. See
\code{min} and \code{max} options for extracting numerical low/high rates.

\subsection{ }{\code{lowest_percentile}, \code{highest_percentile}} Again, special
note should be taken of these methods, which can only be used when rates all
have the same sign (i.e. all negative or all positive). These methods retain
the \code{n}'th lowest or highest percentile of \emph{\strong{absolute}} rate values. That
is, if rates are all negative, \code{method = 'highest_percentile'} will retain
the highest \code{n}'th percentile of rates regardless of sign, that is the \emph{most
negative}. Essentially, these options ignore the sign of the rate. Here \code{n}
should be a percentile value between 0 and 1. For example, to extract the
lowest 10th percentile of rates regardless of sign, you would enter \verb{method = 'lowest_percentile', n = 0.1}.

\subsection{ }{\code{min}, \code{max}} In contrast to \code{lowest} and \code{highest}, these
are \emph{strictly numerical} options which take full account of the sign of the
rate, and can be used where rates are a mix of positive and negative. That
is, \code{method = 'min'} will retain the minimum value numerical rates,
including the most negative. \code{n} indicates how many of the min/max rates to
extract.

\subsection{ }{\code{min_percentile}, \code{max_percentile}} Like \code{min} and \code{max}
these are \emph{strictly numerical} inputs which retain the \code{n}'th minimum or
maximum percentile of the rates and take full account of the sign. Here \code{n}
should be a percentile value between 0 and 1. For example, if rates are all
negative (i.e. typical uptake rates), to extract the lowest 10th percentile
of rates, you would enter \verb{method = 'max_percentile', n = 0.1}. This is
because the \emph{lowest} rates are the \emph{highest numerically} (i.e. least
negative).

\subsection{ }{\code{rate}, \code{rsq}, \code{row}, \code{time}} These methods work on the
respective columns of the \verb{$summary} data frame. For these, \code{n} should be a
numeric vecctor of two values. Matching rows which fall within that range
(inclusive) are returned. For example, to retain only rows where rate is
between 0.05 and 0.08: \verb{method = 'rate', n = c(0.05, 0.08)}. To retain all
rates with a R-Squared above 0.90: \verb{method = 'rsq', n = c(0.9, 1)}. The
\code{row} and \code{time} ranges refer to the original data source and can be used to
constrain results to rates from particular regions of the data. Note, \code{time}
is not the same as \code{duration} - see next section.

\subsection{ }{\code{duration}} This method allows filtering of rates which occur
over a duration range. Here, \code{n} should be a numeric vector of two values
indicating the duration range you are interested in retaining. Use this to
set minimum and maximum durations in the time units of the original data.
For example, \code{n = c(0,500)}, will retain only rates determined over a
maximum of 500 time units. To retain rates over a minimum duration, set this
using the minimum value plus the maximum duration (or simply infinity, e.g.
\code{n = c(500,Inf)}).

\subsection{ }{\code{manual}} This method simply allows particular rows of the
\verb{$summary} data frame to be manually selected to be retained. For example,
to keep only the first result: \verb{method = 'manual', n = 1}. To keep multiple
columns use regular \code{R} selection syntax: \code{n = 1:3}, \code{n = c(1,2,3)}, \code{n = c(5,8,10)}, etc. No value of \code{n} should exceed the number of rows in the
\verb{$summary} data frame.
}
\examples{
 ar_obj <- inspect(intermittent.rd, plot = FALSE) \%>\%
 auto_rate(plot = FALSE)
 ar_filt_neg <- filter_rate(ar_obj, method = "negative", plot = FALSE)
}
