% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowthrough.rd}
\alias{flowthrough.rd}
\title{Flowthrough respirometry data}
\format{A data frame object consisting of and 935 rows (approx 16 mins of data),and
4 columns: Time, oxygen inflow and outflow concentrations, and oxygen delta
(the difference between inflow and outflow concentrations) .}
\usage{
flowthrough.rd
}
\description{
A single experiment on the chiton species \emph{Mopalia lignosa} in a
custom-built flowthrough respirometry system.
Conducted at University of British Columbia, Vancouver, BC, Canada.
}
\details{
Dissolved oxygen units: mg/L
Time units: seconds
Flow rate (L/s): 0.000039
Inflow oxygen concentration (mean, mg/L): 8.88
Specimen ash-free dry mass (kg): 0.000070
}
\author{
Nicholas Carey
}
\keyword{datasets}
