% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowthrough_mult.rd}
\alias{flowthrough_mult.rd}
\title{Multi-column flowthrough respirometry data}
\format{
A data frame object consisting of 3740 rows (approx 62 mins of
data),and 11 columns: time (col 1), oxygen outflow concentrations (cols
2,3,4), inflow concentrations (cols 5,6,7, each paired with outflow
columns), delta concentrations (cols 8,9,10, difference between inflow and
outflow concentrations), and inflow concentrations recorded in a shared
header tank (col 11).
}
\usage{
flowthrough_mult.rd
}
\description{
A simulated dataset for testing and demonstrating flowthrough respirometry
analyses. Contains one column of numeric time data (s), three columns of
outflow oxygen concentrations (oxy.out), three columns of inflow oxygen
concentrations (oxy.in), three columns of delta oxygen concentrations
(oxy.delta, simply the difference between paired columns of inflow and
outflow), and one column of inflow oxygen concentrations recorded from a
shared header tank (oxy.header). The third dataset (col 4) has a period of
higher rates at around the 2000s timepoint, where the specimen increases its
activity then slowly recovers to routine respiration levels.
}
\details{
\itemize{
\item Dissolved oxygen units: mg/L - Time units: seconds
}
}
\author{
Nicholas Carey
}
\keyword{datasets}
