% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{focusplot}
\alias{focusplot}
\title{Generate timeseries plot, focused}
\usage{
focusplot(dt, full = FALSE)
}
\arguments{
\item{dt}{object of class \code{data.frame}.}

\item{full}{logical. Default is FALSE. When set to TRUE, peform a full plot.}
}
\description{
Generate timeseries plot, focused
}
\keyword{internal}
