% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Parse date-time data to numeric}
\usage{
format_data(x, time_format = "ymd_hms", start = 0, event_names = NULL,
  event_times = NULL)
}
\arguments{
\item{x}{vector or data frame containing strings or class POSIX.ct date-time data
to be converted to numeric. If a data frame, assumes these data are in column 1.}

\item{time_format}{string. Code describing structure of date-time data. See details.
Directly relates to functions in the package \code{lubridate}}

\item{start}{numeric. Default = 0. At what time (in seconds) should the formatted time
data start?}

\item{event_names.}{String. Names events during the experiment at the times in
\code{event_times}. E.g. "First flush", "Water speed to 5 cm/s".}

\item{event_times.}{String or class POSIX.ct date/time data matching the events in
\code{event_names}}
}
\value{
A vector or data frame, depending on input.
}
\description{
A function to parse class POSIX.ct or text strings of date-time data to numeric time
for use in respR functions.
}
\details{
** Date-Time data inputs **
Input can be a vector, or data frame. If data frame, assumes date-times are in column 1.
Date-time data can be unspaced or separated by any combination of spaces, forward slashes,
hyphens, dots, colons, semicolons, or underscores. E.g. all these are parsed the same:
"2010-02-28 13:10:23", "20100228 131023", "2010/02/28 13.10;23", "2010 02 28 13_10-23".

Times can be in 24H or 12H with AM/PM.
E.g. "2010-02-28 13:10:23" or "2010-02-28 1:10:23 PM"

AM/PM take precedence over 24H formatting for 01-12h.
E.g. "1:10:23 PM" and "01:10:23 PM" are both same as "13:10:23"

However 24H formatting for 13-24h takes take over AM/PM
E.g. "13:10:23 AM" is identified as "1:10:23 PM" or "13:10:23"

Regardless of input, all data are parsed to numeric time data in seconds duration from
the first entry, unless a \code{start} number is specified in which case the series starts at
that number and all subsequent times are shifted forward by the same amount.

** Date-Time formatting syntax **

\code{time_format} directly corresponds to functions in \code{lubridate}

Time format considerations
\itemize{
\item Should be formatted as date (in lowercase y, m, d) and time (h, m, s) separated by
an underscore.
\item E.g. "2010-02-28 13:10:23" would be ymd_hms
\item Year-Month-Day can be in any order (ymd, ydm, dmy, etc.).
\item Hours-minutes-seconds not required, but if present must be in correct order:
hms, hm, or h
\item Time data without date can be supplied (format = "hms", "hm", "h"), but must not
cross midnight, or dates are required.. It should be formatted as 24h time, or if
longer than 12h have AM/PM appended.
}

Not implemented yet
\code{event_names} and \code{event_times} are so events (e.g. flushes, changes in conditions)
associated with original times in the experiment can be easily associated with new
time data.
}
