% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{fullplot}
\alias{fullplot}
\title{Generate full plot}
\usage{
fullplot(dta, dtb, full = FALSE, interval = NULL)
}
\arguments{
\item{dta}{object of class \code{data.frame}. Full.}

\item{dtb}{object of class \code{data.frame}. Subset.}

\item{full}{logical. Default is FALSE. When set to TRUE, peform a full plot.}

\item{interval}{defaults to NULL. If set to TRUE, will plot interval lines.}
}
\description{
Generate full plot
}
\keyword{internal}
