% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrit.R
\name{generate_mrdf}
\alias{generate_mrdf}
\title{Generate a DO ~ PO2 data table from a DO timeseries}
\usage{
generate_mrdf(dt, width)
}
\description{
Generate a DO ~ PO2 data table from a DO timeseries
}
\keyword{internal}
