% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{get_dec}
\alias{get_dec}
\title{Identify decimal character
For European formatted files with commas as decimal separator}
\usage{
get_dec(path)
}
\description{
Identify decimal character
For European formatted files with commas as decimal separator
}
\keyword{internal}
