% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funs.R
\name{hasNA}
\alias{hasNA}
\title{Check for NA and NaN}
\usage{
hasNA(df, col = 1, summary = F)
}
\arguments{
\item{df}{Dataframe.}

\item{col}{Numeric. Selects column index to parse.}

\item{summary}{Logical. Defaults to FALSE.}
}
\description{
Returns TRUE or FALSE. Also prints an index when summary = T.
Dataframe MUST be parsed through prepData() before this.
}
\examples{
hasNA
}
\author{
Januar Harianto
}
