% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.mo2.R
\name{id.unit}
\alias{id.unit}
\title{Identify units in a character string for conversions.}
\usage{
id.unit(string, cond)
}
\arguments{
\item{string}{character string.}

\item{cond}{string. \code{"input"} or \code{"output"}.}
}
\value{
A character string.
}
\description{
Checks strings and match them to a known database. The database is highly specific - and is divided into "input" or "output" groups.
}
\details{
This is an internal function. The typical user should not see this.
}
