% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.ft.R
\name{inspect.ft}
\alias{inspect.ft}
\title{Explore and visualise flowthrough respirometry data and check for errors}
\usage{
inspect.ft(
  df,
  time = NULL,
  out.o2 = NULL,
  in.o2 = NULL,
  in.o2.value = NULL,
  delta.o2 = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{df}{any object of class \code{data.frame} containing columns of \code{time} and
\code{out.o2} or \code{delta.o2} concentrations, and optionally \code{in.o2}.}

\item{time}{numeric integer. Defaults to 1. Specifies the column number of
the time data.}

\item{out.o2}{numeric value or vector of integers. Defaults to NULL.
Specifies the column number(s) of the outflow O2 data.}

\item{in.o2}{numeric value or vector of integers. Defaults to NULL. Specifies
the column number(s) of the inflow O2 data.}

\item{in.o2.value}{numeric value. Defaults to NULL. If there is no continuous
\code{in.o2} data in \code{df}, this specifies a fixed value of oxygen concentration
or pressure for inflowing water in same units as \code{out.o2}, and is used to
calculate a \code{delta.o2}.}

\item{delta.o2}{numeric vector of integers. Specifies the column number(s) of
the delta O2 data, for data where the user has already calculated the
difference between inflow and outflow oxygen. If entered, \code{out.o2} and
\code{in.o2} should be NULL. Defaults to NULL if no other inputs given.}

\item{plot}{logical. Defaults to TRUE. Plots the data. See Details.}

\item{...}{Allows additional plotting controls to be passed, such as \code{legend = FALSE} and \code{pos}.}
}
\value{
A list object of class \code{inspect.ft}.
}
\description{
\code{inspect.ft} is a data exploration and preparation function that visualises
flowthrough respirometry data, scans it for common issues, and prepares it
for use in other functions in \code{respR}, such as \code{\link[=calc_rate.ft]{calc_rate.ft()}}.
}
\details{
\code{inspect.ft} is specific to \emph{flowthrough} respirometry data. In flowthrough
respirometry (sometimes known as 'open flow' 'open system' or 'continuous
flow' respirometry) rather than calculating a rate from a declining or
increasing oxygen concentration recording in a sealed chamber, instead the
difference (i.e. 'delta') between the inflow and outflow concentrations of a
respirometer receiving water at a constant flow rate is used to calculate an
oxygen consumption or production rate, typically after it has reached a
steady state. Therefore, in general, regions of stable oxygen concentrations
in the outflow recording, indicating stable oxygen delta values, are of
interest. \code{inspect.ft} inspects, visualises, and prepares the data for use in
\code{\link[=calc_rate.ft]{calc_rate.ft()}}. By specifying data types in this function and saving the
output, they do not need to be specified in \code{\link[=calc_rate.ft]{calc_rate.ft()}}.

\code{inspect.ft} requires at least two data inputs; a single column of numeric
\code{time} data, with \emph{either} a column of paired \code{out.o2} concentrations (i.e.
the exhalent or 'downstream' concentrations), \emph{or} a column of already
calculated paired \code{delta.o2} values, that is the difference between outflow
and inflow concentrations, or the outflow concentration corrected by a
background recording from a 'blank' or empty chamber. If an \code{out.o2} column
has been specified, in order to calculate the oxygen delta (and therefore a
rate in \code{\link[=calc_rate.ft]{calc_rate.ft()}}) there must also be an inflow oxygen concentration
input (i.e. the inhalent or 'upstream' concentration). This can either be a
column of paired \code{in.o2} concentrations, or alternatively a single
\code{in.o2.value} value. If a column, the paired values of \code{out.o2} and \code{in.o2}
are used to calculate the oxygen \code{delta.o2}, which is saved in the output,
and used to determine a rate in \code{\link[=calc_rate.ft]{calc_rate.ft()}}. If however the inflow
oxygen concentration is a known, generally unvarying concentration (such as
fully air-saturated water from a header tank) this can be entered as a single
value via \code{in.o2.value} and this is used to calculate the \code{delta.o2}.
}
\section{Data processing and validation checks}{
 \code{inspect.ft} subsets and
classifies the specified columns into a new \code{list} object that can be used
in subsequent functions, reducing the need for additional inputs. Note, use
of \code{inspect.ft} to prepare data for the subsequent functions is optional.
All functions in \code{respR} can accept regular `R`` data objects including
data frames, data tables, tibbles, vectors, etc. Its purpose is quality
control, data visualisation and exploration to assist users in exploring
and preparing their data prior to analysis.

Given an input data frame (\code{df}), the function scans the columns specified
via the \code{time}, \code{out.o2}, \code{in.o2} or \code{delta.o2} columns. If no columns are
specified, by default the functions assumes the first column is \code{time}, and
all others are oxygen data. However, while these columns are checked and
plotted, they are not designated as \code{out.o2}, \code{in.o2} or \code{delta.o2} so if
the object is saved, these column types will have to be specified in
\code{calc_rate.ft}.

The \code{time} column is checked for missing (NA/NaN) values, that values are
sequential, that there are no duplicate times, and that it is numerically
evenly-spaced. Oxygen columns are simply checked for NA/NaN data. See
\strong{Failed Checks} section for what it means for analyses if these checks
produce warnings. The function produces a \code{list} object which, if saved,
can be directly loaded (or \verb{\%>\%} piped) into \code{\link[=calc_rate.ft]{calc_rate.ft()}} for further
analysis.
}

\section{Plot details}{
 A plot of the data is also produced (unless \code{plot = FALSE}), depending on the inputs:
\itemize{
\item a single \code{out.o2} column: a two panel plot. The top plot is both outflow
(black points) and inflow (grey points) oxygen against both time (bottom,
blue axis) and row index (top, red axis). The bottom plot is a rolling
oxygen consumption or production rate, that is the oxygen delta between
inflow and outflow oxygen.
\item a single \code{delta.o2} column: a one panel plot of a rolling oxygen
consumption or production rate, that is the oxygen delta input.
\item multiple \code{out.o2} or \code{delta.o2} columns: a grid plot of all \code{delta.o2}
columns (either as entered or calculated) only. Specific column sets can be
examined as a single plot by using the \code{pos} input. See examples.
\item unspecified columns: all columns are plotted assuming \code{time} in in column
1, and all others are oxygen data. Y-axes are not equal.
}

In rate plots, that is those plotting \code{delta.o2} values, either directly
entered or calculated, regions where rates are stable and consistent will
be horizontally flat.

\emph{\strong{Important:}} Rates in the \code{delta.o2} plots are plotted strictly
numerically on the y-axis. That is, \emph{higher} oxygen consumption rates,
which in \code{respR} are negative since they represent a negative slope of
oxygen against time, will be \emph{lower} on these plots. Higher oxygen
production rates, which are positive, will be higher on the plot.

If the legend obscures part of the plots, it can be suppressed via \code{legend = FALSE} in either the \code{inspect.ft} call, or when using \code{plot()} on the
output object.
}

\section{Multiple data columns}{
 For a quick overview of larger experiments,
multiple columns of \code{out.o2}, \code{in.o2} and \code{delta.o2} can be inspected, but
must share the same numeric time data column specified by the \code{time} input.
Note, multiple column inspection is chiefly intended as a data exploration
step for larger datasets. Subsequent functions such as \code{\link[=calc_rate.ft]{calc_rate.ft()}}
will by default use only the first specified \code{out.o2}, \code{in.o2} and
\code{delta.o2} columns for calculating rates.

If multiple \code{out.o2} columns are specified, \code{in.o2} can be a single column
(if for example all chambers are supplied from the same header tank), in
which case it is used to calculate an oxygen delta for all \code{out.o2}
columns. There can also be multiple \code{in.o2} columns, in which case it is
assumed each \code{out.o2} column is paired with each \code{in.o2} at the same
position, and used to calculate the oxygen \code{delta.o2}. Therefore, they must
have the same number of columns. A single \code{in.o2.value} value in the same
units as \code{out.o2} can also be specified.

Regardless of which inputs are used, multiple columns results in a plot of
each \code{delta.o2} time series. All data are plotted on the same axis range of
both time and oxygen (total range of data). This is chiefly exploratory
functionality to give a quick overview of a dataset, and it should be noted
that while the output \code{list} object will contain all data columns in its
\verb{$dataframe} and \verb{$input_data} elements, subsequent functions such as
\code{\link[=calc_rate.ft]{calc_rate.ft()}}, will by default only use the first inspected data set
(\code{time}, and the first \code{delta.o2}) to calculate rates. Other columns of
oxygen data contained in \code{inspect.ft} output objects can only be used in
these functions by calling the specific output object elements directly.
Best practice is to inspect and assign each individual experiment as
separate \code{inspect.ft} objects. See examples.
}

\section{Failed Checks}{
 It should be noted the data checks in \code{inspect.ft}
are mainly for exploratory purposes; they help diagnose and flag potential
issues with the data. For instance, very long experiments could have had
sensor dropouts the user is completely unaware of. Others are not really
issues at all. For instance, an uneven time warning can result from using
decimalised minutes as the time metric, which happens to be numerically
unevenly spaced. As an additional check, if uneven time is found, the
minimum and maximum intervals in the time data are in the console output,
so a user can see immediately if there are large gaps in the data.

If some of these checks fail, it should \emph{generally} not hinder analysis of
the data. \code{respR} has been coded to rely on linear regressions on exact
data values, and not make assumptions about data spacing or order.
Therefore issues such as missing or NA/NaN values, duplicate or
non-sequential time values, or uneven time spacing should not cause any
erroneous results, as long as they are genuine datapoints and do not occur
over large regions of the data. \code{inspect.ft} however outputs locations (row
numbers) of where these issues occur, allowing users to amend or examine
them before analysis. We would recommend that to be completely confident in
any results from analysis of such data, these issues be considered before
proceeding.
}

\examples{

# inspect.ft flowthrough data
x <- inspect.ft(flowthrough.rd, 1, c(2,3))
x

## pos example
}
