% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data.R
\name{inspect_data}
\alias{inspect_data}
\title{Check for common errors in respirometry data}
\usage{
inspect_data(df, time = NULL, oxygen = NULL, inflow.o2 = NULL,
  outflow.o2 = NULL, highlight = TRUE, plot = TRUE)
}
\arguments{
\item{df}{data frame. Accepts data frame object of any size.}

\item{time}{numeric. Defaults to NULL. This is the time data.}

\item{oxygen}{numaric. Defaults to NULL. This is the dissolved oxygen data.}

\item{inflow.o2}{numeric. Defaults to NULL. This is the incurrent oxygen data.}

\item{outflow.o2}{numeric. Defaults to NULL. This is the excurrent oxygen data.}

\item{highlight}{logical. Defaults to TRUE. Prints location (row #) of errors
detected by the function.}

\item{plot}{logical. Defaults to TRUE. Produces plots for quick visual
diagnostics.}
}
\value{
A list object of class \code{adjust_rate}.
}
\description{
\code{inspect_data()} scans a data frame for specific errors that may affect the
use of functions in \code{respR}. Data checks include:
}
\details{
\itemize{
\item A test for NA/NaN inputs.
\item A test for numeric data.
\item A test for sequential time data.
\item A test for duplicate time data.
\item A test for evenly-spaced time data.
}

Once data checks are complete, the function produces a list object which may
be directly loaded into \code{\link[=calc_rate]{calc_rate()}}, \code{\link[=calc_rate.bg]{calc_rate.bg()}}, \code{\link[=auto_rate]{auto_rate()}} and
\code{\link[=pcrit]{pcrit()}} (NC - i don't think this last one works) for further analyses.
}
\examples{
inspect_data(sardine.rd)
inspect_data(urchins.rd, 1, 5, highlight = FALSE)
}
