% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_help.R
\name{installation_help}
\alias{installation_help}
\title{Guide to installing and using respR v1.1.1}
\usage{
installation_help()
}
\description{
This function opens the help file for this guide.
}
\details{
This help file is a resource to enable publications which have submitted code
using respR v1.1.1 and earlier to have their code remain reproducible.

It contains list of respR dependencies and the most recent versions tested
with and confirmed to work with v1.1.1, and suggestions about how to install
them. More recent versions will likely work with v1.1.1, but in case they
don't the earlier versions can be installed.

If you need to run respR v1.1.1 code, it is probably easiest to install
v1.1.1 and the following dependencies on a base installation of R.

If you have trouble installing respR v1.1.1 (which if you are reading this is
unlikely) you can install it using the following code.

\code{devtools::install_github("januarharianto/respR@v1.1.1")}

Alternatively, download the package source code for v1.1.1 from Github:

https://github.com/januarharianto/respR/releases/tag/v1.1.1

or from Zenodo:

https://zenodo.org/record/6112437

These zip archives, once unzipped, can be set up as an RStudio project and
compiled locally as an R Package.

Alternatively, the \code{R} folder in those links contains the functions. They can
also be run to create the functions locally.
\subsection{Dependencies}{

These are the v1.1.1 dependencies. The package depends on these and will fail
to work correctly without them. Later versions may (and probably will for
many years) work fine, but these are the last versions tested and known to
work with respR v.1.1.1.

The most recent versions can be installed through the usual methods, and will
be installed automatically by installing respR. If they however cause issues,
or you want to ensure the code is being run as it was intended in v1.1.1,
uninstall them and install the older version.

You can install a specific version of a package by adapting the following
code:

\code{devtools::install_version("lubridate", version = "1.8.0", repos = "http://cran.us.r-project.org")}

You can also install older versions from downloaded package zip archives.
Visit https://cran.r-project.org/src/contrib/Archive/ to view and download
the archived versions of CRAN hosted packages.

Dependencies in v1.1.1 (most recent version tested and confirmed to work):

broom 0.7.12, data.table 1.14.2, dplyr 1.0.8, ggplot2 3.3.5, graphics 4.1.2,
grDevices 4.1.2, gridExtra 2.3, lubridate 1.8.0, magrittr 2.0.2, marelac
2.1.10, methods 4.1.2, parallel 4.1.2, readxl 1.3.1, roll 1.1.6, segmented
1.4-0, stats 4.1.2, stringr 1.4.0, utils 4.1.2, xml2 1.3.3

Suggests:

These are packages which may not strictly be necessary, but are suggested to
be installed.

knitr 1.37, rmarkdown 2.11, rMR 1.1.0, DiagrammeR 1.0.8, FishResp 1.1.0,
respirometry 1.3.0, testthat 3.1.2
}
}
