% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intermittent.rd}
\alias{intermittent.rd}
\title{Respirometry data of the sea urchin, \emph{Heliocidaris Erythrogramma}}
\format{A data frame object consisting of 2 columns (time and dissolved oxygen) and 4831 rows.}
\usage{
intermittent.rd
}
\description{
Multiple measurements of oxygen consumption in a single sea urchin, \emph{Heliocidaris erythrogramma}, obtained using intermittent flow respirometry. The experiment was conducted at the Sydney Institute of Marine Science in Sydney, Australia. There are a total of 3 "loops" where the respirometry chamber was flushed. Data was collected using a Vernier Optical DO probe (ODO-BTA).
}
\author{
Nicholas Carey
}
\keyword{datasets}
