% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{is.df.insp.crbg}
\alias{is.df.insp.crbg}
\title{Validate that an input is data.frame, inspect or calc_rate.bg object}
\usage{
is.df.insp.crbg(x)
}
\description{
Validate that an input is data.frame, inspect or calc_rate.bg object
}
\keyword{internal}
