% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{is.nv.cr.ar}
\alias{is.nv.cr.ar}
\title{Validate that an input is numeric vector longer than 1 or calc_rate or auto_rate object}
\usage{
is.nv.cr.ar(x)
}
\description{
Validate that an input is numeric vector longer than 1 or calc_rate or auto_rate object
}
\keyword{internal}
