% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{is.nv.crbg}
\alias{is.nv.crbg}
\title{Validate that an input is numeric vector longer than 1 or calc_rate.bg object}
\usage{
is.nv.crbg(x)
}
\description{
Validate that an input is numeric vector longer than 1 or calc_rate.bg object
}
\keyword{internal}
