% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{is.sn.crbg}
\alias{is.sn.crbg}
\title{Validate that an input is a single numeric value or calc_rate.bg object}
\usage{
is.sn.crbg(x)
}
\description{
Validate that an input is a single numeric value or calc_rate.bg object
}
\keyword{internal}
