% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.rate.R
\name{k.peaks}
\alias{k.peaks}
\title{Perform kernel density estimation of the data and identify the peaks}
\usage{
k.peaks(x)
}
\arguments{
\item{x}{data frame. Makes sense only if it this was an output of \code{rollfit()}.}
}
\value{
A list of 2 objects, "density" and "peaks".
}
\description{
This is an internal function. The typical user should not see this.
}
\keyword{internal}
