% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{kde_fit}
\alias{kde_fit}
\title{Kernel density estimation and fitting}
\usage{
kde_fit(dt, width, by, use = "all")
}
\arguments{
\item{dt}{data.frame object.}

\item{width}{numeric.}

\item{by}{string. time or row}
}
\value{
A list object
}
\description{
Kernel density estimation and fitting
}
\keyword{internal}
