% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{kernel_method}
\alias{kernel_method}
\title{Kernel density function}
\usage{
kernel_method(dt, width, top_only = FALSE)
}
\arguments{
\item{dt}{data.frame object.}

\item{width}{numeric.}

\item{top_only}{logical. Should only top ranked result be returned?}
}
\value{
A list object
}
\description{
This is an internal function for \code{auto_rate()}
}
\keyword{internal}
