% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.rate.R
\name{lmfit}
\alias{lmfit}
\title{Perform OLS regression on subset and save b0, b1 and rsq.}
\usage{
lmfit(x)
}
\arguments{
\item{x}{data frame.}
}
\value{
a data frame.
}
\description{
Used within `calc.rate()`.
}
\details{
This is an internal function. The typical user should not see this.
}
