% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.rate.R
\name{locate.subdfs}
\alias{locate.subdfs}
\title{Subset the main data based on the "by" argument.}
\usage{
locate.subdfs(df, index, by)
}
\arguments{
\item{df}{data frame.}

\item{index}{data frame.}

\item{by}{string.}
}
\value{
A list object.
}
\description{
Also saves row index data in the output.
}
\details{
This is an internal function. The typical user should not see this.
}
