% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.rate.R
\name{match.data}
\alias{match.data}
\title{Match peaks to original data frame}
\usage{
match.data(df, fits, pks, width, bg)
}
\arguments{
\item{df}{data frame.}

\item{fits}{data frame.}

\item{pks}{numeric vector.}

\item{width}{numeric.}

\item{bg}{numeric.}
}
\value{
a data frame.
}
\description{
This is an internal function. The typical user should not see this.
}
\details{
This function uses the the peak values and bin width of the kernel density estimate to match the results to the data in the original data frame.
}
\keyword{internal}
