% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.ft.R
\name{mean.adjust_rate.ft}
\alias{mean.adjust_rate.ft}
\title{Average adjust_rate.ft rates}
\usage{
\method{mean}{adjust_rate.ft}(x, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Average adjust_rate.ft rates
}
