% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{mean.adjust_rate}
\alias{mean.adjust_rate}
\title{Average adjust_rate object rates}
\usage{
\method{mean}{adjust_rate}(x, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Average adjust_rate object rates
}
