% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.int.R
\name{mean.auto_rate.int}
\alias{mean.auto_rate.int}
\title{Average auto_rate.int object rates}
\usage{
\method{mean}{auto_rate.int}(x, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{auto_rate.int object}

\item{pos}{integer(s). Which replicate rates to average.}

\item{export}{logical. Export averaged rate as single value.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Average auto_rate.int object rates
}
\keyword{internal}
