% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{mean.auto_rate}
\alias{mean.auto_rate}
\title{Average auto_rate object rates}
\usage{
\method{mean}{auto_rate}(x, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Average auto_rate object rates
}
