% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.bg.R
\name{mean.calc_rate.bg}
\alias{mean.calc_rate.bg}
\title{Average calc_rate.bg object rates}
\usage{
\method{mean}{calc_rate.bg}(x, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Average calc_rate.bg object rates
}
