% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.ft.R
\name{mean.calc_rate.ft}
\alias{mean.calc_rate.ft}
\title{Average calc_rate.ft object rates}
\usage{
\method{mean}{calc_rate.ft}(x, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Average calc_rate.ft object rates
}
