% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.R
\name{mean.calc_rate}
\alias{mean.calc_rate}
\title{Average calc_rate object rates}
\usage{
\method{mean}{calc_rate}(x, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Average calc_rate object rates
}
