% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_DO.R
\name{mean.convert_DO}
\alias{mean.convert_DO}
\title{Average convert_DO object values}
\usage{
\method{mean}{convert_DO}(x, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Average convert_DO object values
}
