% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.ft.R
\name{mean.convert_rate.ft}
\alias{mean.convert_rate.ft}
\title{Average convert_rate.ft object rates}
\usage{
\method{mean}{convert_rate.ft}(x, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Average convert_rate.ft object rates
}
