% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.R
\name{mean.convert_rate}
\alias{mean.convert_rate}
\title{Average convert_rate object rates}
\usage{
\method{mean}{convert_rate}(x, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Average convert_rate object rates
}
