% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.ft.R
\name{mean.inspect.ft}
\alias{mean.inspect.ft}
\title{Average inspect.ft object rates}
\usage{
\method{mean}{inspect.ft}(x, ...)
}
\arguments{
\item{x}{calc_rate.bg object}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Average inspect.ft object rates
}
