% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{mean.inspect}
\alias{mean.inspect}
\title{Average inspect object rates}
\usage{
\method{mean}{inspect}(x, ...)
}
\arguments{
\item{x}{inspect object}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Average inspect object rates
}
\keyword{internal}
