% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_val.R
\name{method.val}
\alias{method.val}
\title{Validate adjust_rate or auto_rate or oxy_crit method input}
\usage{
method.val(method, source = "adjust_rate")
}
\description{
Validate adjust_rate or auto_rate or oxy_crit method input
}
\keyword{internal}
