% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{midpt}
\alias{midpt}
\title{Get midpoint between two values or within a vector
For getting midpoint timestamps from Time data}
\usage{
midpt(p)
}
\description{
Get midpoint between two values or within a vector
For getting midpoint timestamps from Time data
}
\keyword{internal}
