% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingRegression.R
\name{movingReg}
\alias{movingReg}
\title{Beginner implementation of a moving regression}
\usage{
movingReg(df, span = 0.1)
}
\arguments{
\item{df}{Dataframe. Needs to be formatted by prepData().}

\item{span}{Numeric. Width of window, as a proportion of the length of the
entire dataset. Defaults to 0.1. Must be between 0 and 1.}
}
\description{
Runs multiple regressions of width (span * length of data)
across the entire dataset from beginning to end. Extremely fast
implementation - can run 10k regressions in 3 seconds on a laptop (eyeball
benchmark). You can easily plot the regressions. I still need to figure out
how to pluck the relevant datasets associated with each regression. With some
luck, I may use this to determine basal and max RO2?
}
\examples{
data("sardine")                 # load data
plot(sardine)                   # preview dataset
reg <- movingReg(sardine, .2)   # perform rolling regression
plot(reg$beta)                  # plot change in rate over index of time
                                # max and min rates can be visualised
plot(density(reg$beta))         # density plot of rate
}
\author{
Januar Harianto
}
