% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{multi.p}
\alias{multi.p}
\title{Plot the main data, and the subset data, in the same window.}
\usage{
multi.p(df, sdf, title = T)
}
\arguments{
\item{df}{data frame.}

\item{sdf}{data frame.}

\item{title}{logical.}
}
\value{
A plot.
}
\description{
Plot the main data, and the subset data, in the same window.
}
