% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_funs.R
\name{nainf.omit}
\alias{nainf.omit}
\title{Omit NA, NaN, Inf and -Inf from a vector or dataframe columns}
\usage{
nainf.omit(x)
}
\description{
For using with, for example, range to get axis range values in 'inspect'.
Previously, na.omit was used, then discovered data files with Inf values.
This causes axis limit range to be Inf, and xlim/ylim don't accept infinite
axes!
}
\details{
If x is dataframe, it returns a vector of all columns appended together.
Only useful for getting range in this case, don't use for anything else.
}
\keyword{internal}
