% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{overlap.ft.p}
\alias{overlap.ft.p}
\title{Plot convert_rate.ft summary tables}
\usage{
overlap.ft.p(
  x,
  highlight = NULL,
  pos = NULL,
  legend = TRUE,
  quiet = FALSE,
  rate.rev = TRUE,
  msg = "overlap.ft.p",
  ...
)
}
\arguments{
\item{x}{\code{convert_rate.ft}, \code{convert_rate.ft_select} object}

\item{highlight}{integer. Which result in the summary table to highlight on
the plots. Defaults to 1. If it is outside the range of the summary rows it
will default to 1.}

\item{pos}{integer(s). Choose which summary table rows to plot.}

\item{legend}{logical. Suppress plot legends.}

\item{quiet}{logical. Suppress console output.}

\item{msg}{string. For adding custom text to start of messages.}

\item{...}{Allows additional plotting controls to be passed.}
}
\value{
A plot of the auto_rate object results
}
\description{
Plots \code{convert_rate.ft} summary table regressions in a way that
visualises how they are positioned within the data timeseries.
}
\keyword{internal}
