% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy_crit.R
\name{oxy_crit}
\alias{oxy_crit}
\title{Calculate critical oxygen tensions}
\usage{
oxy_crit(
  x,
  time = NULL,
  oxygen = NULL,
  rate = NULL,
  width = 0.1,
  plot = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{x}{data frame or object of class \code{inspect}. This is the data to
analyse.}

\item{time}{numeric vector. Defaults to NULL. This specifies the column
number of the time data.}

\item{oxygen}{numeric vector. Defaults to NULL. This specifies the column
number(s) of the oxygen data.}

\item{rate}{numeric vector. Defaults to NULL. This specifies the column
number(s) of rate data.}

\item{width}{numeric. Determines the width of the rolling regression used to
determine rolling rate. Proportion of the length of data. Defaults to 0.1,
or 10\% of total rows.}

\item{plot}{logical. Defaults to TRUE.}

\item{parallel}{logical. Defaults to FALSE. Should parallel processing be
used?}
}
\value{
A list object of class \code{oxy_crit}.
}
\description{
A function to calculate the critical oxygen tension, or the O2 concentration
below which uptake rate becomes dependent upon oxygen concentration. It is
calculated by both the "broken-stick" regression (BSR) approach, adopted from
Yeager and Ultsch (1989), and the segmented regression approach, presented by
Muggeo (2003).
}
\details{
The default data input is an \code{inspect} object, or \code{data.frame} containing
time~dissolved oxygen data (e.g. \code{squid.rd}). Columns of each can be
specified, the default being they are the first two columns (i.e. \code{time = 1},
\code{oxygen = 2}). If an \code{inspect} object is used the data frame is extracted
automatically and column identifiers are not required since these were
already identified and extracted in those functions.

To calculate Pcrit, the function requires data in the form of oxygen uptake
rate against dissolved oxygen (DO) concentration. The function performs a
rolling regression on time~O2 data to determine rates, and matches these
against a rolling mean of the DO data. The function then performs the two
\eqn{P_{crit}}{P[crit]} analyses methods on these data. The width of the
rolling regression is determined by the \code{width} argument. In most cases, the
default width (10\% of the data length) works well, but this may vary with
data that has abrupt changes in rate, or is particularly noisy.

Alternatively, existing rate~DO data may be used, with the \code{rate} input
identifying the column with the rate data. In this case the function performs
the two \eqn{P_{crit}}{P[crit]} analyses on these data directly without any
processing.
}
\examples{
# Intensive, and a large dataset, so may take some time.
\dontrun{
data("squid.rd")
pcrit(squid.rd)
}
}
\references{
Yeager DP, Ultsch GR (1989) Physiological regulation and
conformation: A BASIC program for the determination of critical points.
Physiological Zoology 62:888–907. doi: 10.1086/physzool.62.4.30157935

Muggeo V (2008) Segmented: an R package to fit regression models with
broken-line relationships. R News 8:20–25.
}
