% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrit2.R
\name{pcrit2}
\alias{pcrit2}
\title{Calculate critical oxygen tension, \eqn{P_{crit}}{P[crit]}}
\usage{
pcrit2(df, width = floor(0.1 * nrow(df)), has.rate = F, plot = T)
}
\arguments{
\item{df}{data frame.}

\item{width}{numeric.}

\item{has.rate}{logical.}

\item{plot}{logical.}
}
\value{
A list object.
}
\description{
Critical oxygen tension can be calculated either by the "broken-stick" regression (BSR) approach, adopted from Yeager and Ultsch (1989), or the segmented regression approach, presented by Muggeo (2003).
}
\details{
To determine P_crit, RO_2 (or MO_2) and ambient oxygen concentration (PO_2) must be provided. If RO_2 has not been calculated, pcrit() can automatically perform a rolling regression directly from raw respirometry data to generate the data and arrange it against PO_2, and then use one of the two methods to estimate the value.
}
\examples{
NULL
}
\author{
Januar Harianto & Nicholas Carey
}
