% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.ft.R
\name{plot.adjust_rate.ft}
\alias{plot.adjust_rate.ft}
\title{Plot adjust_rate.ft objects}
\usage{
\method{plot}{adjust_rate.ft}(x, ...)
}
\arguments{
\item{x}{adjust_rate.ft object}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot adjust_rate.ft objects
}
\keyword{internal}
