% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{plot.adjust_rate}
\alias{plot.adjust_rate}
\title{Plot adjust_rate objects}
\usage{
\method{plot}{adjust_rate}(x, ...)
}
\arguments{
\item{x}{calc_rate.bg object}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot adjust_rate objects
}
\keyword{internal}
