% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.rate.R
\name{plot.auto.rate}
\alias{plot.auto.rate}
\title{Plot an object of class \code{auto.rate}}
\usage{
\method{plot}{auto.rate}(x, rank = 1)
}
\arguments{
\item{x}{object of class \code{auto.rate}.}

\item{rank}{numeric. Defaults to 1. Changing this number will allow user to view other ranked results.}
}
\value{
a plot output.
}
\description{
Plot an object of class \code{auto.rate}
}
