% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{plot.auto_rate}
\alias{plot.auto_rate}
\title{Plot auto_rate objects}
\usage{
\method{plot}{auto_rate}(
  x,
  pos = 1,
  panel = FALSE,
  quiet = FALSE,
  legend = TRUE,
  rate.rev = TRUE,
  ...
)
}
\value{
A plot. No returned value.
}
\description{
Plot auto_rate objects
}
