% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.bg.R
\name{plot.calc_rate.bg}
\alias{plot.calc_rate.bg}
\title{Plot calc_rate.bg objects}
\usage{
\method{plot}{calc_rate.bg}(x, pos = NULL, quiet = FALSE, legend = TRUE, ...)
}
\value{
A plot. No returned value.
}
\description{
Plot calc_rate.bg objects
}
