% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.ft.R
\name{plot.calc_rate.ft}
\alias{plot.calc_rate.ft}
\title{Plot calc_rate.ft objects}
\usage{
\method{plot}{calc_rate.ft}(x, pos = NULL, quiet = FALSE, legend = TRUE, rate.rev = TRUE, ...)
}
\value{
A plot. No returned value.
}
\description{
Plot calc_rate.ft objects
}
