% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.int.R
\name{plot.calc_rate.int}
\alias{plot.calc_rate.int}
\title{Plot calc_rate.int objects}
\usage{
\method{plot}{calc_rate.int}(x, pos = NULL, quiet = FALSE, legend = FALSE, type = "rep", ...)
}
\arguments{
\item{x}{calc_rate.int object}

\item{pos}{integer. Which replicate(s) to plot up to a maximum of 20.
Defaults to 1:20.}

\item{quiet}{logical. Suppress console output.}

\item{legend}{logical. Suppress labels and legends.}

\item{type}{logical. Type of plot to produce. \code{"rep"}, \code{"full"}, or
\code{"cr"}. Defaults to \code{"rep"}. See Details.}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot calc_rate.int objects
}
\keyword{internal}
