% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.R
\name{plot.calc_rate}
\alias{plot.calc_rate}
\title{Plot calc_rate objects}
\usage{
\method{plot}{calc_rate}(x, pos = 1, quiet = FALSE, panel = NULL, legend = TRUE, ...)
}
\value{
A plot. No returned value.
}
\description{
Plot calc_rate objects
}
