% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRate.R
\name{plot.calcRate}
\alias{plot.calcRate}
\title{Plot for class calcRate}
\usage{
\method{plot}{calcRate}(x, ...)
}
\arguments{
\item{x}{An object of class 'calcRate'.}

\item{...}{Additional arguments may be passed into the function.}
}
\value{
ggplot2 objects, arranged in grids
}
\description{
Plot for class calcRate
}
\examples{
# Load data
data(sardine)
sardine.full <- calcRate(sardine) # if you want to run the entire dataset
plot(sardine.full)

}
