% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_DO.R
\name{plot.convert_DO}
\alias{plot.convert_DO}
\title{Plot convert_DO objects}
\usage{
\method{plot}{convert_DO}(x, ...)
}
\arguments{
\item{x}{convert_DO object}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot convert_DO objects
}
