% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.ft.R
\name{plot.inspect.ft}
\alias{plot.inspect.ft}
\title{Plot inspect.ft objects}
\usage{
\method{plot}{inspect.ft}(
  x,
  width = NULL,
  pos = NULL,
  quiet = FALSE,
  legend = TRUE,
  rate.rev = TRUE,
  ...
)
}
\value{
A plot. No returned value.
}
\description{
Plot inspect.ft objects
}
