% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{plot.inspect}
\alias{plot.inspect}
\title{Plot inspect objects}
\usage{
\method{plot}{inspect}(
  x,
  width = NULL,
  pos = NULL,
  quiet = FALSE,
  legend = TRUE,
  rate.rev = TRUE,
  ...
)
}
\value{
A plot. No returned value.
}
\description{
Plot inspect objects
}
