% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy_crit.R
\name{plot.oxy_crit}
\alias{plot.oxy_crit}
\title{Plot oxy_crit objects}
\usage{
\method{plot}{oxy_crit}(x, legend = TRUE, quiet = FALSE, panel = NULL, rate.rev = TRUE, ...)
}
\value{
A plot. No returned value.
}
\description{
Plot oxy_crit objects
}
