% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_lin.R
\name{plot.test_lin}
\alias{plot.test_lin}
\title{Plot output of \code{test_lin()}}
\usage{
\method{plot}{test_lin}(x, show = c("all", "a", "b", "c", "d"), ...)
}
\description{
This is an internal generic. Produces 4 plots.
}
\keyword{internal}
