% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{plot_ar}
\alias{plot_ar}
\title{Plot auto_rate summary tables}
\usage{
plot_ar(x, highlight = NULL, pos = NULL, legend = TRUE, ...)
}
\arguments{
\item{x}{\code{auto_rate} or \code{auto_rate_subset} object}

\item{highlight}{integer. Which result in the summary table to highlight on
the plots. Defaults to 1. If it is outside the range of the \code{pos} input it
will be shown on the top plot, but will not be visible on the bottom plot.}

\item{pos}{integer(s). What range of original summary table rows to plot in
lower plot. Defaults to all.}

\item{legend}{logical. Suppress plot legends.}

\item{...}{Allows additional plotting controls to be passed.}
}
\value{
A plot of the auto_rate object results
}
\description{
Plots \code{auto_rate} summary table regressions in a way that visualises how they
are positioned within the data timeseries. If it is an \code{auto_rate_subset}
object, it will plot the subset regressions using the ranks of the original
results, so you can compare the subset and original.
}
