% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{plot_ar_grid}
\alias{plot_ar_grid}
\title{Plots multiple auto_rate results in a nice way
using base plot
x = auto_rate or auto_rate_subset object}
\usage{
plot_ar_grid(
  x,
  pos = NULL,
  msg = "subset_rate",
  title = "Subset",
  max_msg = "",
  ...
)
}
\description{
Plots multiple auto_rate results in a nice way
using base plot
x = auto_rate or auto_rate_subset object
}
\keyword{internal}
