% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{plot_multi_ar}
\alias{plot_multi_ar}
\title{Plots multiple auto_rate results in a nice way
x = auto_rate or auto_rate_filt object
n = max no. of plots
THIS IS SLOOOOOOOW
Will probably have to revert to base plotting}
\usage{
plot_multi_ar(x, n = 9)
}
\description{
Plots multiple auto_rate results in a nice way
x = auto_rate or auto_rate_filt object
n = max no. of plots
THIS IS SLOOOOOOOW
Will probably have to revert to base plotting
}
\keyword{internal}
