% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.rate.R
\name{print.auto.rate}
\alias{print.auto.rate}
\title{Print an object of class \code{auto.rate}}
\usage{
\method{print}{auto.rate}(x, rank = 1)
}
\arguments{
\item{x}{object of class \code{auto.rate}.}

\item{rank}{numeric. Defaults to 1. Changing this number will allow user to view other ranked results.}
}
\value{
a print output.
}
\description{
Print an object of class \code{auto.rate}
}
