% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{print.auto_rate}
\alias{print.auto_rate}
\title{Print auto_rate objects}
\usage{
\method{print}{auto_rate}(x, pos = 1, ...)
}
\value{
Print to console. No returned value.
}
\description{
Print auto_rate objects
}
