% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.mo2.R
\name{print.calc.mo2}
\alias{print.calc.mo2}
\title{Print output with rank argument}
\usage{
\method{print}{calc.mo2}(x, rank = 1)
}
\arguments{
\item{x}{object of rank `calc.mo2`.}

\item{rank}{numeric.}
}
\description{
Print output with rank argument
}
