% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.bg.R
\name{print.calc_rate.bg}
\alias{print.calc_rate.bg}
\title{Print calc_rate.bg objects}
\usage{
\method{print}{calc_rate.bg}(x, ...)
}
\arguments{
\item{x}{calc_rate.bg object}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Print calc_rate.bg objects
}
\keyword{internal}
