% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.ft.R
\name{print.calc_rate.ft}
\alias{print.calc_rate.ft}
\title{Print calc_rate.ft objects}
\usage{
\method{print}{calc_rate.ft}(x, pos = 1, ...)
}
\value{
Print to console. No returned value.
}
\description{
Print calc_rate.ft objects
}
