% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.int.R
\name{print.calc_rate.int}
\alias{print.calc_rate.int}
\title{Print calc_rate.int objects}
\usage{
\method{print}{calc_rate.int}(x, pos = NULL, ...)
}
\arguments{
\item{x}{calc_rate.int object}

\item{pos}{integer. Which replicate to print.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Print calc_rate.int objects
}
\keyword{internal}
