% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.R
\name{print.calc_rate}
\alias{print.calc_rate}
\title{Print calc_rate objects}
\usage{
\method{print}{calc_rate}(x, pos = 1, ...)
}
\arguments{
\item{x}{calc_rate object}

\item{pos}{integer. Which result to print.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Print calc_rate objects
}
