% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_DO.R
\name{print.convert_DO}
\alias{print.convert_DO}
\title{Print convert_DO objects}
\usage{
\method{print}{convert_DO}(x, ...)
}
\arguments{
\item{x}{convert_DO object}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Print convert_DO objects
}
\keyword{internal}
