% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.ft.R
\name{print.convert_rate.ft}
\alias{print.convert_rate.ft}
\title{Print convert_rate.ft objects}
\usage{
\method{print}{convert_rate.ft}(x, pos = NULL, ...)
}
\value{
Print to console. No returned value.
}
\description{
Print convert_rate.ft objects
}
