% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.R
\name{print.convert_rate}
\alias{print.convert_rate}
\title{Print convert_rate objects}
\usage{
\method{print}{convert_rate}(x, pos = 1, ...)
}
\arguments{
\item{x}{convert_rate object}

\item{pos}{integer. Which result to print.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Print convert_rate objects
}
\keyword{internal}
