% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.ft.R
\name{print.inspect.ft}
\alias{print.inspect.ft}
\title{Print inspect.ft objects}
\usage{
\method{print}{inspect.ft}(x, ...)
}
\value{
Print to console. No returned value.
}
\description{
Print inspect.ft objects
}
