% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{print.inspect}
\alias{print.inspect}
\title{Print inspect objects}
\usage{
\method{print}{inspect}(x, ...)
}
\arguments{
\item{x}{inspect object}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Print inspect objects
}
\keyword{internal}
