% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy_crit.R
\name{print.oxy_crit}
\alias{print.oxy_crit}
\title{Print oxy_crit objects}
\usage{
\method{print}{oxy_crit}(x, ...)
}
\arguments{
\item{x}{oxy_crit object}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Print oxy_crit objects
}
\keyword{internal}
