% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{print_dens}
\alias{print_dens}
\title{Prints the density object for summary.auto_rate S3
Basically copied from stats:::print.density and edited to make
it more compact}
\usage{
print_dens(x, digits = NULL, ...)
}
\arguments{
\item{x}{auto_rate object.}

\item{digits}{numeric.}
}
\value{
Console output
}
\description{
Prints the density object for summary.auto_rate S3
Basically copied from stats:::print.density and edited to make
it more compact
}
\keyword{internal}
