% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{qqplot}
\alias{qqplot}
\title{Generate QQ plot}
\usage{
qqplot(model, full = FALSE)
}
\arguments{
\item{model}{an \code{lm()} object.}

\item{full}{logical. Default is FALSE. When set to TRUE, peform a full plot.}
}
\description{
Generate QQ plot
}
\keyword{internal}
