% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respR.R
\docType{package}
\name{respR-package}
\alias{respR}
\alias{respR-package}
\title{respR: Import, Process, Analyse, and Calculate Rates from Respirometry Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a structural, reproducible workflow for the processing and analysis of respirometry data. It contains analytical functions and utilities for working with oxygen time-series to determine respiration or oxygen production rates, and to make it easier to report and share analyses. See Harianto et al. 2019 \doi{10.1111/2041-210X.13162}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/januarharianto/respr}
  \item \url{https://januarharianto.github.io/respR/}
  \item \doi{10.1111/2041-210X.13162}
}

}
\author{
\strong{Maintainer}: Nicholas Carey \email{nicholascarey@gmail.com}

Authors:
\itemize{
  \item Nicholas Carey \email{nicholascarey@gmail.com}
  \item Januar Harianto \email{januar.harianto@sydney.edu.au}
}

}
\keyword{internal}
