% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.rate.R
\name{rollfit}
\alias{rollfit}
\title{Perform rolling regressions of specific width}
\usage{
rollfit(df, width)
}
\arguments{
\item{df}{data frame. The data to perform rolling regressions on.}

\item{width}{numeric. The rolling window.}
}
\value{
A data frame of coefficients.
}
\description{
This is an internal function. The typical user should not see this.
}
\keyword{internal}
