% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{rolling_reg}
\alias{rolling_reg}
\title{Rolling regression}
\usage{
rolling_reg(dt, width, by)
}
\arguments{
\item{width}{numeric. width in number of rows or time}

\item{by}{string. time or row}

\item{df}{data.frame object.}
}
\value{
a data.table object
}
\description{
Rolling regression
}
\keyword{internal}
