% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{rolling_reg_row}
\alias{rolling_reg_row}
\title{Perform regular rolling regression}
\usage{
rolling_reg_row(df, width)
}
\arguments{
\item{df}{data.frame object.}

\item{width}{numeric.}
}
\value{
a data.table object
}
\description{
Perform regular rolling regression
}
\keyword{internal}
