% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{rolling_reg_time}
\alias{rolling_reg_time}
\title{Perform rolling regression based on time units.}
\usage{
rolling_reg_time(df, width)
}
\arguments{
\item{df}{data.frame object.}

\item{width}{numeric.}
}
\value{
a data.table object
}
\description{
Perform rolling regression based on time units.
}
\keyword{internal}
