% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{rollplot}
\alias{rollplot}
\title{Title}
\usage{
rollplot(obj, pos = 1, full = FALSE)
}
\arguments{
\item{obj}{object of class \code{auto_rate}.}

\item{pos}{numeric. Ranked result.}

\item{full}{logical. Default is FALSE. When set to TRUE, peform a full plot.}
}
\description{
Title
}
\keyword{internal}
