% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.rate.R
\name{scale.rate}
\alias{scale.rate}
\title{Title}
\usage{
\method{scale}{rate}(x, o2.unit = NULL, time.unit = NULL,
  output.unit = NULL, volume = NULL, mass = NULL, S = 35, t = 25,
  P = 1.013253)
}
\arguments{
\item{P}{}
}
\description{
Title
}
