% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.mo2.R
\name{scale.unit}
\alias{scale.unit}
\title{Scale units of the same base.}
\usage{
\method{scale}{unit}(b1, unit.in, unit.out)
}
\arguments{
\item{b1}{numeric vector.}

\item{unit.in}{character string.}

\item{unit.out}{character string.}
}
\value{
numeric vector.
}
\description{
This function performs simple conversion in units with the same base, e.g. x/s to x/h, or mmol/x to umol/x.
}
\details{
This is an internal function. The typical user should not see this.
}
