% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_smooth_func.R
\name{stat_smooth_func}
\alias{stat_smooth_func}
\title{Print equation to line}
\usage{
stat_smooth_func(mapping = NULL, data = NULL, geom = "smooth",
  position = "identity", ..., method = "auto", formula = y ~ x,
  se = TRUE, n = 80, span = 0.75, fullrange = FALSE,
  level = 0.95, method.args = list(), na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, xpos = NULL, ypos = NULL)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with aes or aes_string}

\item{data}{Data input}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlappling points on this layer}

\item{...}{other arguments}

\item{method}{smoothing method (function) to use, eg. lm, glm, gam, loess, rlm}

\item{formula}{formula to use in smoothing function}

\item{se}{display confidence interval around smooth?}

\item{n}{number of points to evaluate smoother at}

\item{span}{span}

\item{fullrange}{should the fit span the full range of the plot, or just the data}

\item{level}{level of confidence interval to use (0.95 by default)}

\item{method.args}{methods to use}

\item{na.rm}{remove NAs?}

\item{show.legend}{show legend?}

\item{inherit.aes}{inherit aes?}

\item{xpos}{x-coordinate location of text}

\item{ypos}{y-coordinate location of text}
}
\description{
Taken from https://gist.github.com/kdauria/524eade46135f6348140
Credit: https://stackoverflow.com/a/27959418
Internal function for plotting.
}
\keyword{internal}
