% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{static_roll}
\alias{static_roll}
\title{Normal rolling regression}
\usage{
static_roll(df, win)
}
\arguments{
\item{df}{data.frame object.}

\item{win}{numeric. width in number of rows}
}
\value{
a data.table object
}
\description{
This is an internal function.
}
\keyword{internal}
