% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeUnits.R
\name{storeUnits}
\alias{storeUnits}
\title{Define units and experimental conditions for the current session}
\usage{
storeUnits(Time, O2, Vol, Mass)
}
\arguments{
\item{Time}{Character.}

\item{O2}{Character.}

\item{Vol}{Character.}

\item{Mass}{Character.}
}
\description{
This function is run first to establish measurement units that
will be used for all calculations in subsequent functions.
}
\examples{
storeUnits()
storeUnits('min', 'mgl-1', 'l', 'kg')
}
