% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_val.R
\name{StP.check}
\alias{StP.check}
\title{Checks if an oxygen concentration/pressure unit or metabolic rate unit
requires temperature, salinity, and pressure to convert to another unit.}
\usage{
StP.check(unit, type)
}
\description{
\code{type} should be \code{"oxy"} or \code{"mr"}.
}
\details{
Returns \code{TRUE} if it requires t,S,P or \code{FALSE} if not.
}
\keyword{internal}
