% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_val.R
\name{StP.val}
\alias{StP.val}
\title{Returns an error message if an oxygen concentration/pressure unit or
metabolic rate unit requires S or t and either is NULL, or a message if it
requires P and it is NULL.}
\usage{
StP.val(unit, type, S, t, P, P.chk = TRUE, msg)
}
\description{
\code{type} should be \code{"oxy"} or \code{"mr"}.
\code{P.chk} - perform P out of normal range check. Only want to do this once per function.
}
\details{
Also returns default value of P = 1.013253 if it is NULL.
}
\keyword{internal}
