% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.bg.R
\name{subset.data}
\alias{subset.data}
\title{Truncate a data frame to create a subset of the data}
\usage{
\method{subset}{data}(x, from, to, by)
}
\arguments{
\item{x}{data frame.}

\item{from}{numeric.}

\item{to}{numeric.}

\item{by}{string. "time", "row", "o2" or "proportion".}
}
\value{
A \code{data.table} object.
}
\description{
This is an internal function. This function extracts a subset data frame
based on a given set of rules.
}
\keyword{internal}
