% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.ft.R
\name{summary.adjust_rate.ft}
\alias{summary.adjust_rate.ft}
\title{Summarise adjust_rate.ft objects}
\usage{
\method{summary}{adjust_rate.ft}(object, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{object}{adjust_rate.ft object}

\item{pos}{integer(s). Which summary row(s) to print.}

\item{export}{logical. Export summary table as data frame.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise adjust_rate.ft objects
}
