% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{summary.adjust_rate}
\alias{summary.adjust_rate}
\title{Summarise adjust_rate objects}
\usage{
\method{summary}{adjust_rate}(object, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Summarise adjust_rate objects
}
