% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.rate.R
\name{summary.auto.rate}
\alias{summary.auto.rate}
\title{Summarise an object of class \code{auto.rate}}
\usage{
\method{summary}{auto.rate}(x, n = 5)
}
\arguments{
\item{x}{object of class \code{auto.rate}.}

\item{n}{numeric. Defaults to 5. The number of results to show.}
}
\value{
a summary output.
}
\description{
Summarise an object of class \code{auto.rate}
}
\keyword{internal}
