% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.int.R
\name{summary.auto_rate.int}
\alias{summary.auto_rate.int}
\title{Summarise auto_rate.int objects}
\usage{
\method{summary}{auto_rate.int}(object, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{object}{auto_rate.int object}

\item{pos}{integer(s). Which replicate(s) to print.}

\item{export}{logical. Export summary table as data frame.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise auto_rate.int objects
}
\keyword{internal}
