% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{summary.auto_rate}
\alias{summary.auto_rate}
\title{Summarise auto_rate objects}
\usage{
\method{summary}{auto_rate}(object, pos = NULL, export = FALSE, print.kds = TRUE, ...)
}
\arguments{
\item{object}{auto_rate object}

\item{pos}{integer(s). Which summary row(s) to print.}

\item{export}{logical. Export summary table as data frame.}

\item{print.kds}{logical. Print summary of the kernel density analysis for
'linear' method.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise auto_rate objects
}
