% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.bg.R
\name{summary.calc_rate.bg}
\alias{summary.calc_rate.bg}
\title{Summarise calc_rate.bg objects}
\usage{
\method{summary}{calc_rate.bg}(object, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Summarise calc_rate.bg objects
}
