% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.int.R
\name{summary.calc_rate.int}
\alias{summary.calc_rate.int}
\title{Summarise calc_rate.int objects}
\usage{
\method{summary}{calc_rate.int}(object, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{object}{calc_rate.int object}

\item{pos}{integer(s). Which replicate(s) to print.}

\item{export}{logical. Export summary table as data frame.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise calc_rate.int objects
}
\keyword{internal}
