% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.R
\name{summary.calc_rate}
\alias{summary.calc_rate}
\title{Summarise calc_rate objects}
\usage{
\method{summary}{calc_rate}(object, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Summarise calc_rate objects
}
