% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_DO.R
\name{summary.convert_DO}
\alias{summary.convert_DO}
\title{Summarise convert_DO objects}
\usage{
\method{summary}{convert_DO}(object, ...)
}
\arguments{
\item{object}{convert_DO object}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise convert_DO objects
}
