% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.ft.R
\name{summary.convert_rate.ft}
\alias{summary.convert_rate.ft}
\title{Summarise convert_rate.ft objects}
\usage{
\method{summary}{convert_rate.ft}(object, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Summarise convert_rate.ft objects
}
