% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.R
\name{summary.convert_rate}
\alias{summary.convert_rate}
\title{Summarise convert_rate objects}
\usage{
\method{summary}{convert_rate}(object, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Summarise convert_rate objects
}
