% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.ft.R
\name{summary.inspect.ft}
\alias{summary.inspect.ft}
\title{Summarise inspect.ft objects}
\usage{
\method{summary}{inspect.ft}(object, ...)
}
\arguments{
\item{object}{inspect.ft object}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise inspect.ft objects
}
\keyword{internal}
