% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{summary.inspect}
\alias{summary.inspect}
\title{Summarise inspect objects}
\usage{
\method{summary}{inspect}(object, ...)
}
\value{
Print to console. No returned value.
}
\description{
Summarise inspect objects
}
