% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_lin_data}
\alias{test_lin_data}
\title{Output objects for the function \code{test_lin()}}
\format{List of multiple output objects of class \code{test_lin}.}
\usage{
test_lin_data
}
\description{
This data contains the results of 9 separate performance checks
on \code{auto_rate()}'s linear detection algorithm (i.e. \code{method = "linear"}).
These test results are used to assess and discuss the performance of
\code{auto_rate()} in our online vignette.
}
\author{
Januar Harianto
}
\keyword{datasets}
