% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{time_lm}
\alias{time_lm}
\title{Subset data by time and perform a linear regression.}
\usage{
time_lm(df, start, end)
}
\description{
This is an internal function. Used with \code{\link[=time_roll]{time_roll()}} and \code{\link[=auto_rate]{auto_rate()}}.
}
\keyword{internal}
