% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.args.R
\name{unit.args}
\alias{unit.args}
\title{Print examples for unit arguments in \code{calc.mo2()}}
\usage{
unit.args()
}
\value{
A list.
}
\description{
Print examples for unit arguments in \code{calc.mo2()}
}
\examples{
unit.args()

}
