% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_val.R
\name{unit_type_o1}
\alias{unit_type_o1}
\title{Search for and classify oxygen units as used in concentration or rates}
\usage{
unit_type_o1(unit, msg = "")
}
\description{
Search for and classify oxygen units as used in concentration or rates
}
\keyword{internal}
