% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_val.R
\name{units.clean}
\alias{units.clean}
\title{Cleans units from units.val to remove the suffix (.o2, .flow, etc)}
\usage{
\method{units}{clean}(unit, is)
}
\description{
These are the 'clean' or parsed names we want to use in outputs. All input
units get parsed to these.
}
\details{
Could make code much simpler with a regex for everything after the "." but i
like the specificity of this.
}
\keyword{internal}
