% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_val.R
\name{units.val}
\alias{units.val}
\title{Check unit string against allowed values. See util_fns.R file for regex
patterns}
\usage{
\method{units}{val}(unit, is, msg = "units.val")
}
\description{
These names (before the .o2, .vol, etc.) are the 'clean' or parsed names we
want to use in outputs. All inputs units get parsed to these in units.clean
}
\keyword{internal}
