% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_val.R
\name{units_val}
\alias{units_val}
\title{Check unit string against allowed values. See util_fns.R file for regex
patterns}
\usage{
units_val(unit, is, msg = "units_val")
}
\description{
These names (before the .o2, .vol, etc.) are the 'clean' or parsed names we
want to use in outputs. All inputs units get parsed to these in units_clean
}
\keyword{internal}
