% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{val_meth}
\alias{val_meth}
\title{Validate adjust_rate method input}
\usage{
val_meth(method)
}
\description{
Validate adjust_rate method input
}
\keyword{create}
\keyword{for}
\keyword{internal}
\keyword{linear/exp}
\keyword{logical}
\keyword{methods}
