% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate_funs.R
\name{validate_auto_rate}
\alias{validate_auto_rate}
\title{Validation function for auto_rate}
\usage{
validate_auto_rate(df, by, method)
}
\arguments{
\item{df}{data.frame object.}

\item{by}{string.}

\item{method}{string.}
}
\value{
a data.frame object of the original data
}
\description{
This is an internal function for \code{auto_rate()}. Used to validate inputs
}
\keyword{internal}
