% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeb_intermittent.rd}
\alias{zeb_intermittent.rd}
\title{Respirometry data of a zebrafish, \emph{Danio rerio}}
\format{
A data frame object consisting of 2 columns (time and dissolved
oxygen) and 79251 rows (approx 22h of data).
}
\usage{
zeb_intermittent.rd
}
\description{
Multiple measurements (106 replicates, plus initial and end background
measurements) of oxygen consumption in a zebrafish, \emph{Danio rerio}, obtained
using intermittent flow respirometry. Data kindly provided by Davide
Thambithurai (University of Glasgow). Note, the data has been injected with
random noise, and volume and mass below are not the actual values from the
experiment, but for example only.
}
\details{
Dissolved oxygen units: mg/L
\itemize{
\item Time units: seconds
\item Chamber volume (L): 0.1
\item Specimen wet mass (kg): 0.001
}

Replicate structure (Rows - Experiment section)
\itemize{
\item 1:4999 - Start background recording
\item 5000:5839 - First replicate for MMR (14 mins duration)
\item 5840:75139 - 105 further replicates of 11 minutes duration each (660 rows)
\item 75140:79251 - End background recording
}
}
\author{
Davide Thambithurai
}
\keyword{datasets}
