% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.ft.R
\name{adjust_rate.ft}
\alias{adjust_rate.ft}
\title{Adjust rates in flowthrough respirometry to account for background
respiration or oxygen flux.}
\usage{
adjust_rate.ft(x, by)
}
\arguments{
\item{x}{numeric. A single numeric value, numeric vector, or object of class
\code{calc_rate.ft}. This is the experimental rate value(s) to be adjusted.}

\item{by}{numeric. A numeric value, numeric vector or object of class
\code{calc_rate.ft}. This contains the background rate used to perform the
adjustment to \code{x}. If the vector or \code{calc_rate.ft} contains multiple rates,
they will be averaged to produce a single adjustment value.}
}
\value{
Output is a list object of class \code{adjust_rate.ft}
}
\description{
The \code{adjust_rate.ft} function adjusts an oxygen uptake or production rate
(for example, as determined in \code{\link{calc_rate.ft}}) for background oxygen use by
microbial organisms, or other removal (or input) of oxygen during
\emph{flowthrough} respirometry experiments. The function accepts numeric values,
as well as regular \code{calc_rate.ft} objects.
}
\details{
\emph{\strong{Note:}} take special care with the \strong{sign} of the rate used for
adjustments. In \code{respR} oxygen uptake rates are negative values, as they
represent a negative slope of oxygen against time. Background rates will
normally also be a negative value. See examples.

\code{adjust_rate.ft} allows the rate (or multiple rates) in \code{x} to be adjusted by
the background rate in \code{by}. There are several ways of determining the
background rate, or performing background corrections depending on the setup
of the experiment.

For experiments in which an empty "blank" experiment has been run, and the
background rate generally does not change over the course of the experiment
(that is, the oxygen delta between inflow and outflow concentrations remains
consistent), it is recommended the rate be determined and saved via the
\verb{[inspect.ft]} and \verb{[calc_rate.ft]} functions and then entered as the \code{by}
input as either a value or the saved \code{calc_rate.ft} object. In this case, the
\verb{$rate} element of the \code{calc_rate.ft} object is used to adjust all rates in
\code{x}. If there are multiple rates in \verb{$rate}, the mean value of all rates
found is used. In this way, a single blank experiment can be applied to
numerous actual experiments, or several blank experiments can be averaged to
provide a single average adjustment rate. Alternatively, once blank
experiments have been analysed to produce a background correction value, can
be and this entered via \code{by} as a numeric value.

For experiments in which an empty "blank" experiment has been run alongside
actual experiments in parallel, and background rate may increase or decrease
over time (or there may be other variations such as in the inflow oxygen
concentrations), it is recommended you \emph{NOT} use this function. Instead, the
paired blank oxygen concentration data should be used in \verb{[inspect.ft]} as
the \code{in.o2} input. In this way, the calculated delta oxygen values take
account of whatever background or other variation in oxygen is occurring in
the blank chamber with respect to time.

For all background operations, it is important the same \code{flowrate} is used.

For adjustments, all rates in \code{x}, whether entered as values or as a
\code{calc_rate.ft} object, are adjusted by subtracting the mean of all background
rates in \code{by}. Again, take special note of the \emph{sign} of these rates. See
examples.
\subsection{Output}{

If the \code{x} input is numeric, the output will be a numeric of the same length
containing adjusted rates.

If the \code{x} input is a \code{calc_rate.ft} object, the output will be identical in
structure, but of class \code{adjust_rate.ft} and containing the additional
elements \verb{$adjustment} and \verb{$rate.adjusted} and these also added to summary
metadata. If \code{x} is a value or vector, the output is a list object of class
\code{adjust_rate.ft} containing three elements: \verb{$rate}, \verb{$adjustment}, and
\verb{$rate.adjusted}. The \verb{$rate.adjusted} element will be the one converted when
the object is passed to \code{convert_rate.ft}.
}
}
\examples{
# Note that oxygen uptake rates are negative by default (since it represents a
# decrease in dissolved oxygen and negative slope). Therefore, typically both
# rate and background rate values are negative.

# Simple background respiration correction to a single rate.
# This is (-7.44) - (-0.04) = -7.40
adjust_rate.ft(x = -7.44, by = -0.04)

# Oxygen input adjustment
# This is (-7.44) - (0.1) = -7.54
adjust_rate.ft(x = -7.44, by = 0.1)

# Mean background respiration correction to a single rate.
adjust_rate.ft(x = -7.44, by = c(-0.04, -0.05, -0.06))

# Mean background respiration correction to multiple rates.
out <- adjust_rate.ft(x = c(-7.44, -7.20, -7.67),
                   by = c(-0.04, -0.05, -0.06))
summary(out)

# Paired background respiration correction to multiple rates.
out <- adjust_rate.ft(x = c(-7.44, -7.20, -7.67),
                   by = c(-0.04, -0.05, -0.06))
summary(out)
}
