% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{adjust_rate}
\alias{adjust_rate}
\title{Simple rate adjustment function (background or oxygen flux)}
\usage{
adjust_rate(
  x,
  by,
  by2 = NULL,
  method = "mean",
  t_x = NULL,
  t_by = NULL,
  t_by2 = NULL
)
}
\arguments{
\item{x}{numeric. A single value, or any object of class \code{calc_rate} or \code{auto_rate}.
This object contains the rate value that needs to be adjusted.}

\item{by}{either a numeric, or an object of class \code{calc_rate.bg}. This is the
value that is used to perform the correction/adjustment.}
}
\value{
A list object of class \code{adjust_rate}.
}
\description{
A simple rate correction function. Please note that the \strong{sign} of the
number must be considered in the correction. Background respiration is
normally a negative number, while oxygen flux may or may not be negative.
}
\details{
To account for background respiration and/or oxygen flux in open respirometry
experiments, we have provided a simple function to perform either, or both,
corrections. As a separate function we may develop this to support more
complex correction methods (e.g. non-linear) in the future.
}
\examples{
# Note that respiration is negative by default (since it represents a
# decrease in dissolved oxygen), so both values are negative!
# Background respiration correction
adjust_rate(-7.44, -0.04) # this is (-7.44) - (-0.04) = -7.40
# Oxygen flux correction
adjust_rate(-7.44, 0.1) # this is (-7.44) - (0.1) = -7.54
}
