% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{algae.rd}
\alias{algae.rd}
\title{Oxygen production respirometry data}
\format{
A data frame object consisting of 1200 rows (20 h of data),and
2 columns: time in hours, oxygen in \% air saturation.
}
\usage{
algae.rd
}
\description{
Data from a respirometry experiment on algae which shows oxygen production
over time.
}
\details{
\itemize{
\item Dissolved oxygen units: \% Air Saturation
\item Time units: hours
\item Respirometer volume: 0.1 L
\item Temperature: 12 C
\item Salinity: 30 ppt
}
}
\author{
Nicholas Carey
}
\keyword{datasets}
