% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.rate.R
\name{auto.rate}
\alias{auto.rate}
\title{Determine min, max or 'ideal' rate}
\usage{
auto.rate(df, width = round(0.1 * nrow(df)), logic = "automatic",
  background = NULL)
}
\arguments{
\item{df}{Data frame.}

\item{width}{Numeric.}

\item{logic}{Character.}

\item{background}{Numeric, or an object of class \code{'calc.bg.rate'}}

\item{plot}{Character.}
}
\value{
A summary.
}
\description{
Determine min, max or 'ideal' rate
}
\examples{
TBC
}
