% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{auto_rate}
\alias{auto_rate}
\title{auto.rate version 2}
\usage{
auto_rate(df, width = NULL, by = "time", method = "default",
  plot = TRUE)
}
\arguments{
\item{df}{NULL}

\item{width}{NULL}

\item{by}{NULL}

\item{method}{NULL}
}
\description{
auto.rate version 2
}
\examples{
NULL
}
