% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{auto_rate}
\alias{auto_rate}
\title{Automatically determine rate of change in oxygen concentration over time}
\usage{
auto_rate(df, width = NULL, by = "row", method = "linear",
  plot = TRUE, parallel = FALSE)
}
\arguments{
\item{df}{data frame, or object of class \code{adjust_rate}. This is the data to
process.}

\item{width}{numeric. Width of the rolling regression. Defaults to
\code{floor(0.2*length of data)} if NULL. The length of data can either be time
or row, as defined by the \code{by} argument. If a number is supplied and it is
less than 1, the function automatically applies the equation \code{floor(width * length of data}. Otherwise it will simply use the number as the width.}

\item{by}{string. \code{"row"} or \code{"time"}. Defaults to \code{"row"}. However, if the
function detects an irregular time series, a warning will be issued to
recommend changing this argument to \code{"time"}. In most cases \code{"row"} is used
by default as it is efficient. Switching to \code{"time"} causes the function to
perform checks for irregular time at every iteration of the rolling
regression, even if time is evenly spaced, which adds to computation time.}

\item{method}{string. \code{"linear"}, \code{"max"}, \code{"min"} or \code{"interval"}. Defaults
to \code{linear}. Note that if \code{"linear"} is selected, the argument \code{width} will
be set to default.}

\item{plot}{logical. Defaults to TRUE. Plot the results.}

\item{parallel}{logical. Defaults to TRUE. Should parallel processing be
used?}
}
\value{
A list object of class \code{auto_rate}.
}
\description{
\code{auto_rate} automatically performs a rolling regression on a data frame to
determine the \emph{most linear, maximum, minimum}, or \emph{interval} rate of change
in oxygen concentration over time. First, a rolling regression of specified
\code{width} is performed on the entire dataset to obtain all possible values. The
computations are then ranked (or, arranged), based on the "\code{logic}" argument,
and the output is summarised.
}
\details{
\strong{Units}

There are no units of measurement involved in \code{auto_rate}. This is a
deliberate decision. Units are called in a later function when volumetric
and/or mass-specific rates of oxygen use are computed in \code{\link[=convert_rate]{convert_rate()}} and
\code{\link[=convert_DO]{convert_DO()}}.

\emph{\strong{Ranking algorithms}}

At present, \code{auto_rate()} contains four ranking algorithms that can be called
with the argument \code{method}:
\itemize{
\item \code{linear}: Uses kernel density estimation (KDE) to detect the "most
linear" sections of the timeseries. This is achieved by using the smoothing
bandwidth of the KDE to re-sample the "peaks" in the KDE to determine linear
regions in the data.
\item \code{max}: regressions are arranged from highest values, to the lowest.
\item \code{min}: regressions are arranged from lowest values, to the highest.
\item \code{interval}: non-overlapping regressions are extracted from the rolled
regrssions. They are not ranked.
}
}
\examples{
# most linear section of the entire data
data("flowthrough.rd")
auto_rate(flowthrough.rd, parallel = FALSE)

# LONG EXAMPLES
\dontrun{
data("sardine.rd")
# what is the lowest rate over a 10 minute (600s) period?
auto_rate(sardine.rd, method = "min", width = 600, by = "time", parallel = FALSE)
# what is the highest rate over a 10 minute (600s) period?
auto_rate(sardine.rd, method = "max", width = 600, by = "time", parallel = FALSE)
}
}
