% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.bg.rate.R
\name{calc.bg.rate}
\alias{calc.bg.rate}
\title{Calculate background respiration}
\usage{
calc.bg.rate(df, timecol = 1, bgcol = 2, plot = T)
}
\arguments{
\item{df}{Data frame.}

\item{timecol}{Numeric. Defaults to 1. The column index of the data frame to use as time data.}

\item{bgcol}{Numeric or a list. Defaults to 2. The column index/indices of the data frame to use for DO data. A list e.g. `c(2:5)` will calculate the rate for each column individually, and then average the results.}
}
\value{
A summary list of rates, and the mean rate.
}
\description{
Calculate background respiration
}
\examples{
data(urchin2013)
calc.bg.rate(urchin2013, 1, c(18:19))
# save as object to use in other functions:
urbg <- calc.bg.rate(urchin2013,1,c(18:19))

}
