% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.mo2.R
\name{calc.mo2}
\alias{calc.mo2}
\title{Calculate volume and/or mass-specific rate of change in oxygen concentration}
\usage{
calc.mo2(x, o2.unit = "mg/l", time.unit = "s", output.unit = "mg/s/kg",
  volume = NULL, mass = NULL, rank = 1, S = 35, t = 25,
  P = 1.013253)
}
\arguments{
\item{x}{numeric, or an object of class \code{calc.rate} or \code{auto.rate}. The input object to calculate the conversion.}

\item{o2.unit}{character string. Your unit for oxygen concentration. Defaults to \code{"mg/l"}. Options can be viewed by calling \code{check.units()}.}

\item{time.unit}{character string. Your unit for time. Defaults to \code{"s"}. Options can be viewed by calling \code{check.units()}.}

\item{output.unit}{character string. Your unit for output. Defaults to \code{"mg/s/kg"}. Options can be viewed by calling \code{check.units()}.}

\item{volume}{numeric. The volume of medium to correct to. Must be in litres (L).}

\item{mass}{numeric (optional). The mass of the speciment to correct to. Must be in kilograms (kg).}

\item{rank}{numeric (optional). If the input is of class \code{auto.rate}, the user may convert any of the ranked outputs by specifying the rank here. Note: not used for other input types.}

\item{S}{numeric. Salinity, defaults to 35.}

\item{t}{numeric. Temperature in degrees C, defaults to 25.}

\item{P}{numeric. Pressure unit in bar, defaults to 1.013253.}
}
\value{
An object containing a list of outputs:
\describe{
\item{\code{input.class}}{String. Helps the user identify the source of the input
\code{x}. Possible values: \code{calc.rate}, \code{auto.rate}, \code{numeric}.}
\item{\code{input.id}}{String. If the source of the input \code{x} is \code{auto.rate},
identifies it's ranking method. Possible values: \code{maxmin}, \code{interval},
\code{automatic}, \code{calc.rate}, \code{numeric}.}
\item{\code{volume}}{Numeric. Same as input parameter.}
\item{\code{mass}}{Numeric. Same as input parameter.}
\item{\code{rank}}{Numeric. Same as input parameter.}
\item{\code{salinity}}{Numeric. Same as input parameter.}
\item{\code{temperature}}{Numeric. Same as input parameter.}
\item{\code{pressure}}{Numeric. Same as input parameter.}
\item{\code{summary}}{Data frame. Contains all inputs and converted outputs.}
\item{\code{unit.in}}{String. Input units identified by the function.}
\item{\code{unit.out}}{String. Output units identified by the function.}
\item{\code{converted}}{Numeric (vector). Converted value(s).}
\item{\code{weighted}}{Numeric. The weighted average of conversions (useful for
intermittent data).}
}
}
\description{
This is a conversion function that can convert value(s) of rate of change in oxygen concentration to volume-specific, and/or volume- and mass-specific rate of change in oxygen concentration.
}
\details{
\strong{Integration}

\code{calc.mo2()} integrates well with output objects from \code{\link[=calc.rate]{calc.rate()}} and \code{\link[=auto.rate]{auto.rate()}}. Alternatively, the user may convert any numeric vector.

\strong{String matching}

A string-matching algorithm is used to identify the units for \code{o2.unit}, \code{time.unit} and \code{output.unit}. First of all, the most common styles of writing units are supported. For example, these are all the same: \code{mg/L; mg/l, mg L-1, mgL-1, mg per litre, mg.l-1, mg.L-1}. Similarly for time, these are all the same: \code{s, second, sec}. Therefore, an outpot for weight-specific rate of change in O~2~ concentration can be provided intuitively, e.g. \code{"mg/s/kg"} or \code{"mg s-1 kg-1"}, and the function should accept variations in style without error.

\strong{Support}

Supported units can be viewed by calling the function \code{\link[=unit.args]{unit.args()}}. As of now, we do not support conversions for units of pressure, e.g. mmHg, Torr, but are planning to do so in a future version.
}
\examples{
x <- calc.rate(sardine, bg = 0.0001234)
calc.mo2(x, "\%/s", "mg/s", volume = 1.2)

x <- calc.rate(intermittent, c(200,2300,4100), c(1800,3200,4600), by = 'time')
calc.mo2(x, "mg/l/s", "mg/s/kg", volume = 1, mass = 1)

x <- auto.rate(squid, logic = "interval")
calc.mo2(x, "\%/s", "mg/s/g", volume = 1, mass = .5)

}
\seealso{
\code{\link[=calc.rate]{calc.rate()}}, \code{\link[=auto.rate]{auto.rate()}}, \code{\link[=unit.args]{unit.args()}}
}
\author{
Januar Harianto & Nicholas Carey
}
