% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.bg.R
\name{calc_rate.bg}
\alias{calc_rate.bg}
\title{Calculate background respiration rates}
\usage{
calc_rate.bg(x, time = NULL, oxygen = NULL, plot = TRUE)
}
\arguments{
\item{x}{data frame, \code{inspect} or \code{inspect_data} object. This is the data to
process.}

\item{time}{integer. Defaults to NULL. This specifies the column number of
the time data. If NULL, function assumes time data are in Column 1.}

\item{oxygen}{integer vector. Defaults to NULL. This specifies the column
number(s) of the oxygen data. Multiple columns of oxygen can be specified.
If NULL, function assumes oxygen data are in all columns of the data frame
except Column 1.}

\item{plot}{logical. Defaults to TRUE. Will plot the results for visual
inspection.}
}
\value{
A list object of class \code{calc_bg.rate}.
}
\description{
This function uses simple linear regression to calculate the rate of change
of oxygen over time for background corrections of the main data. The
background data must be in the same time and oxygen units as the data to be
corrected. Multiple columns of background oxygen measures, as long as the
time data are identical between measurements.The function returns rates for
all columns, and also calculates an average rate.
}
\details{
If you need to subset the data to remove regions you don't want to use, see
\code{\link{subset_data}}.

There are no units involved in \code{calc_rate.bg()}. This is a deliberate
decision. Units are called in a later function when volumetric and/or
mass-specific rates of oxygen use are computed in \code{\link[=convert_rate]{convert_rate()}} and
\code{\link[=convert_DO]{convert_DO()}}.
}
\examples{
data("urchins.rd")
calc_rate.bg(urchins.rd, time = 1, oxygen = 18:19)
}
