% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.bg.R
\name{calc_rate.bg}
\alias{calc_rate.bg}
\title{Calculate background respiration rates}
\usage{
calc_rate.bg(x, xcol = 1, ycol = 2, from = NULL, to = NULL,
  by = "time", plot = T)
}
\arguments{
\item{x}{data frame.}

\item{xcol}{numeric. Defaults to \code{1}.}

\item{ycol}{numeric vector. Defaults to \code{2}. Can have length > 1.}

\item{from}{numeric.}

\item{to}{numeric.}

\item{by}{string. "time" or "row". Defaults to "time".}

\item{plot}{logical. Defaults to TRUE.}
}
\value{
A list object of class \code{calc.bg.rate}.
}
\description{
This function uses simple linear regression to automatically calculate the
rate of change of oxygen over time for background corrections of the main
data. Can be used on multiple datasets of background measures, as long as the
time data are identical between measurements.
}
\details{
#' There are no units involved in \code{calc_rate.bg()}. This is a deliberate
decision. Units are called in a later function when volume- and/or
weight-specific rates of oxygen concentration are computed in
\code{\link[=convert.rate]{convert.rate()}} and \code{\link[=scale.rate]{scale.rate()}}.
}
\examples{
calc_rate.bg(urchins.rd, xcol = 1, ycol = 18:19, from = 5, to = 45, by = "time")
}
