% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.ft.R
\name{calc_rate.ft}
\alias{calc_rate.ft}
\title{Calculate rate of change in oxygen in flowthrough respirometry}
\usage{
calc_rate.ft(x = NULL, outflow.o2 = NULL, inflow.o2 = NULL,
  flowrate = NULL, plot = TRUE)
}
\arguments{
\item{x}{data frame or object of class \code{inspect}.}

\item{outflow.o2}{numeric. Defaults to NULL. This specifies the outflow O2
column if a data frame or \code{inspect} object is provided. Otherwise, this is
a numeric vector for outflow oxygen concentration.}

\item{inflow.o2}{numeric. Defaults to NULL. This specifies the inflow O2
column if a data frame or \code{inspect} object is provided. Otherwise, this is
a numeric vector for inflow oxygen concentration.}

\item{flowrate}{numeric vector. The flow rate. Must be in \emph{Litres} per unit
time (s,m,h), for example L/s. The units are not required to be entered in
this function; you will specify them when you perform conversions later in
\code{convert_rate}.}

\item{plot}{logical. Defaults to TRUE. Plots the data.}
}
\value{
An object of class "calc_rate.ft".
}
\description{
Calculates rate of oxygen uptake in flowthrough respirometry given a
flow-rate and both inflow (upstream of the chamber) and outflow (downstream
of the chamber) oxygen concentrations. Can return a single rate value, or
multiple and mean values based on continuous data.
}
\details{
Input can be numeric values or vectors - see examples. In addition, a
\code{data.frame} or \code{inspect} object can be entered. In this case \code{outflow.o2}
and \code{inflow.o2} are indices indicting the columns containing these data.
Note, an \code{inflow.o2} column \emph{must} be present in the data.frame, even if this
is a single fixed value repeated to the correct length. Note, a numeric time
elapsed column can be present, but is not required. In flowthrough data,
uptake rate calculations use O2 delta and the flow rate.

There are no units involved in \code{calc_rate.ft}. This is a deliberate decision.
Units are called in a later function when absolute and/or mass-specific
rates of oxygen use are computed in \code{\link[=convert_rate]{convert_rate()}} and \code{\link[=convert_DO]{convert_DO()}}.
}
\examples{
# Single numeric values
calc_rate.ft(inflow.o2 = 8.88, outflow.o2 = 8.17, flowrate = 0.000039)
# Numeric values and vector
data("flowthrough.rd")
calc_rate.ft(inflow.o2 = 8.88, outflow.o2 = flowthrough.rd$o2.out,
  flowrate = 0.000039)
# Vectors
data("flowthrough.rd")
calc_rate.ft(inflow.o2 = flowthrough.rd$o2.in,
  outflow.o2 = flowthrough.rd$o2.out, flowrate = 0.000039)
# A data frame
data("flowthrough.rd")
calc_rate.ft(flowthrough.rd, outflow.o2 = 2, inflow.o2 = 3, flowrate = 0.00039)
}
