% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.ft.R
\name{calc_rate.ft}
\alias{calc_rate.ft}
\title{Calculate flowthrough rate of change in dissolved oxygen}
\usage{
calc_rate.ft(df = NULL, in.col = 1, out.col = 2, inflow = NULL,
  outflow = NULL, delta = NULL, flowrate = NULL)
}
\arguments{
\item{df}{data frame.}

\item{in.col}{numeric. Used only if df is not NULL. Selects the inflow data
by column number.}

\item{out.col}{numeric. Used only if df is not NULL. Selects the outflow data
by column number.}

\item{inflow}{numeric vector. Incurrent oxygen concentration.}

\item{outflow}{numeric vector. Excurrent oxygen concentration.}

\item{delta}{numeric vector. The O2 differential between inflow and outflow.}

\item{flowrate}{numeric vector. The flow rate. No unit of measurement is
expected, although you will specify it when you perform conversions later
on.}
}
\value{
An object of class "calc_rate.ft".
}
\description{
I need a description here.
}
\examples{

}
