% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRate.R
\name{calcRate}
\alias{calcRate}
\title{Calculate rate over time. Need a better title.}
\usage{
calcRate(df, from = NULL, to = NULL, by = "proportion",
  expt.type = "closed")
}
\arguments{
\item{df}{Data frame.}

\item{from}{Numeric.}

\item{to}{Numeric.}

\item{by}{Character.}

\item{expt.type}{Character. Not yet active.}

\item{...}{Additional parameters can be added here, if needed.}
}
\value{
A list.
}
\description{
Calculate rate over time. Need a better title.
}
\examples{
# Load data
data(sardine)
sardine.full <- calcRate(sardine) # if you want to run the entire dataset
plot(sardine.full)

# if you want to specify range, include from and to arguments.
sardine.subset <- calcRate(sardine, from = 0.7, to = 0.2)
plot(sardine.subset)

}
