% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.R
\name{check.input}
\alias{check.input}
\title{Check data frame for structural errors}
\usage{
check.input(df, xcol = NULL, ycol = NULL, plot = T)
}
\arguments{
\item{df}{data frame. containing at least 2 columns of data.}

\item{xcol}{numeric. Corresponds to column number. \code{xcol = 1} will subset the
first column of the data frame. Must be time data.}

\item{ycol}{numeric. Corresponds to column number. \code{ycol = 2} will subset the
second column of the data frame. Must be dissolved oxygen data.}

\item{plot}{logical. When TRUE, a plot of the data is automatically produced
if the data frame subset contains 2 columns. Otherwise, this does nothing.}
}
\value{
If a 2-column data frame is checked, \code{check.input} returns a data
frame object that can be saved for use in \code{\link[=calc.rate]{calc.rate()}} and \code{\link[=auto.rate]{auto.rate()}}.
}
\description{
\code{check.input} scans a data frame for specific errors that may affect
the use of functions in \code{respr}. Data checks include:
\itemize{
\item A test for NA inputs.
\item A test for NaN inputs.
\item A test for numeric data.
\item A test for duplicates (relevant for time data only).
\item A test for evenly-spaced data (relevant for time data only).
\item A check for column lengths.
}
}
\details{
You may run \code{check.input} on a multi-column data frame. Alternatively, you
may run \code{check.input} on a 2-column data frame and it will automatically
assume that the first column is time data and the second column, oxygen
concentration data. A 2-column data frame can also be extracted directly from
a multi-column data frame if the \code{x} and \code{y} arguments are used within the
function. For 2-column data frames, \code{check.input} will automatically remove
NAs and produce a new data frame object, as long as no critical errors are
found in the data. If critical errors are found (e.g. non numeric data), the
function will stop. Two-column checks are more informative - since it is
assumed that the x column is time data and y column is dissolved oxygen data,
time-specific errors can be called out.
}
\examples{
check.input(urchin2013)  # this does not produce a data frame object
check.input(urchin2013, xcol = 1, ycol = 15)  # this produces a data frame object
check.input(sardine)  # this produces a data frame object

}
\author{
Januar Harianto & Nicholas Carey
}
