% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_val.R
\name{class.val}
\alias{class.val}
\title{Validates acceptable classes of inputs
Set single or multiple inputs to TRUE. If x is any one of these a TRUE
will be returned}
\usage{
class.val(
  x,
  int = FALSE,
  int.sing = FALSE,
  int.mult = FALSE,
  num = FALSE,
  num.sing = FALSE,
  num.mult = FALSE,
  df = FALSE,
  cr = FALSE,
  cr.sing = FALSE,
  cr.mult = FALSE,
  cr.int = FALSE,
  cr.int.sing = FALSE,
  cr.int.mult = FALSE,
  ar = FALSE,
  ar.sing = FALSE,
  ar.mult = FALSE,
  crbg = FALSE,
  crbg.sing = FALSE,
  crbg.mult = FALSE,
  insp = FALSE
)
}
\description{
Validates acceptable classes of inputs
Set single or multiple inputs to TRUE. If x is any one of these a TRUE
will be returned
}
\keyword{internal}
