% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_DO.R
\name{convert_DO}
\alias{convert_DO}
\title{Convert units of dissolved oxygen.}
\usage{
convert_DO(x, from = NULL, to = NULL, S = NULL, t = NULL,
  P = 1.013253)
}
\arguments{
\item{x}{numeric vector or object of class \code{calc_rate}, \code{auto_rate} or
\code{adjust_rate}. This is the value(s) that you want to convert from.}

\item{from}{string. The unit to convert from. See \code{\link[=unit_args]{unit_args()}} for details.}

\item{to}{string. The unit to convert to. See \code{\link[=unit_args]{unit_args()}} for details.}

\item{S}{numeric. Salinity (ppt). Defaults to NULL. Used only in conversion
of some units. See \code{\link[=unit_args]{unit_args()}} for details.}

\item{t}{numeric. Temperature(°C). Defaults to NULL. Used only in conversion
of some units. See \code{\link[=unit_args]{unit_args()}} for details.}

\item{P}{numeric. Pressure (bar). Defaults to 1.013253. Used only in
conversion of some units. See \code{\link[=unit_args]{unit_args()}} for details.}
}
\value{
A list.
}
\description{
This is a conversion function that performs basic conversions between units
of dissolved oxygen in aquatic respirometry. Some units require temperature
(\code{t}), salinity (\code{S}), and atmospheric pressure (\code{P}) to be specified. See
\code{\link[=unit_args]{unit_args()}} for details. For freshwater experiments, salinity should be set
to zero (i.e. S = 0).
}
\examples{
# Perform conversion on an object of class `calc_rate`:
x <- calc_rate(sardine.rd)
convert_DO(x, from = "percent", to = "mg L-1")

# Or, perform on a numeric
convert_DO(100, from = "percent", to = "mg L-1", S = 33, t = 18)

convert_DO(sardine.rd[[2]], from = "percent", to = "torr")
}
