% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_MR.R
\name{convert_MR_print}
\alias{convert_MR_print}
\title{Print results of convert_MR}
\usage{
convert_MR_print(oldrates, newrates, from, to)
}
\arguments{
\item{oldrates}{The original rates}

\item{newrates}{The converted rates}

\item{from}{The original unit}

\item{to}{The converted unit}
}
\description{
Print results of convert_MR
}
\keyword{internal}
