% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.ft.R
\name{convert_rate.ft}
\alias{convert_rate.ft}
\title{Convert a unitless oxygen rate value from flowthrough respirometry to
absolute, mass-specific or area-specific rates}
\usage{
convert_rate.ft(
  x,
  o2.unit = NULL,
  flowrate.unit = NULL,
  output.unit = NULL,
  mass = NULL,
  area = NULL,
  S = NULL,
  t = NULL,
  P = 1.013253
)
}
\arguments{
\item{x}{numeric value or vector, object of class \code{\link[=calc_rate.ft]{calc_rate.ft()}} or
\code{\link[=adjust_rate.ft]{adjust_rate.ft()}}. Contains the rate(s) to be converted.}

\item{o2.unit}{string. The dissolved oxygen unit of the original data used to
determine the rate in \code{x}.}

\item{flowrate.unit}{string. The unit of the flowrate through the
respirometer. See Details.}

\item{output.unit}{string. The output unit to convert the input rate to.
Should be in the correct order: "O2/Time" or "O2/Time/Mass" or
"O2/Time/Area".}

\item{mass}{numeric. Mass/weight in \strong{kg}. This is the mass of the specimen
if you wish to calculate mass-specific rates.}

\item{area}{numeric. Surface area in \strong{m^2}. This is the surface area of the
specimen if you wish to calculate surface area-specific rates.}

\item{S}{numeric. Salinity (ppt). Defaults to NULL. Used in conversion of
some oxygen units. Fresh water should be entered as \code{S = 0}.}

\item{t}{numeric. Temperature(°C). Defaults to NULL. Used in conversion of
some oxygen units.}

\item{P}{numeric. Pressure (bar). Used in conversion of some oxygen units.
Defaults to a standard value of 1.013253 bar.}
}
\description{
\code{convert_rate.ft} converts a unitless rate derived from \code{\link[=calc_rate.ft]{calc_rate.ft()}} or
\code{\link[=adjust_rate.ft]{adjust_rate.ft()}} into an absolute rate (i.e. whole specimen or whole
chamber), mass-specific rate (i.e. normalised by specimen mass), or
area-specific rate (i.e. normalised by specimen surface area) in any common
unit.
}
\details{
By default, \code{convert_rate.ft} converts the \verb{$rate} element from
\code{calc_rate.ft} objects, or the \verb{$rate.adjusted} element from \code{adjust_rate.ft}
objects if these are entered as the \code{x} input. Alternatively, a numeric value
or vector of rates can be input as \code{x}.
\subsection{Units}{

The \code{o2.unit} of the original data used to calculated the rate is required.
Concentration units should use only SI units (\code{L} or \code{kg}) for the
denominator, e.g. \code{"mg/L"}, \code{"mmol/kg"}. Percentage saturation of air or
oxygen is accepted, as are oxygen pressure units. See \code{\link[=unit_args]{unit_args()}} for
details.

An \code{output.unit} is also required. If left NULL, The default of \code{"mgO2/h"} is
used, or \code{"mgO2/h/kg"} or \code{"mgO2/h/m2"} if a \code{mass} or \code{area} respectively
has been entered. The \code{output.unit} must be in the sequence \emph{O2-Time} (e.g.
\code{"mg/h"}) for absolute rates, \emph{O2-Time-Mass} (e.g. \code{"mg/h/kg"}) for
mass-specific rates, and \emph{O2-Time-Area} (e.g. \code{"mg/h/cm2"}) for surface
area-specific rates.

Note, some oxygen input or output units require temperature (\code{t}) and
salinity (\code{S}) to perform conversions. For freshwater experiments, salinity
should be entered as zero (i.e. \code{S = 0}). Strictly speaking these also
require an atmospheric pressure (\code{P}) input. In reality, it has a relatively
minor effect within normal ranges, however the default value of 1.013253 bar
(standard pressure at sea level) can be changed if desired. See \code{\link[=unit_args]{unit_args()}}
for details.

The \code{flowrate.unit} is required and should be the units of the \code{flowrate}
used in \code{calc_rate.ft} to calculate the rate, and should be in the form of
volume (L, ml, or ul) per unit time (s,m,h,d), for example in \code{"L/s"}. Note,
the volume component does \emph{NOT} represent the volume of the respirometer, and
the time component does \emph{NOT} represent the units or recording interval of
the original data.

The function uses an internal database and a fuzzy string matching algorithm
to accept various unit formatting styles. For example, \verb{'mg/l', 'mg/L', 'mgL-1', 'mg l-1', 'mg.l-1'} are all parsed the same. See \code{\link[=unit_args]{unit_args()}} for
details of accepted units and their formatting. See also \code{\link[=convert_val]{convert_val()}}
for simple conversion between units.
}

\subsection{Output}{

Returns a \code{list} object containing the \verb{$rate.input}, and converted rate(s)
in \verb{$rate.output} in the \verb{$output.unit}, as well as inputs and summary
elements.
}

\subsection{S3 Generic Functions}{

Saved output objects can be used in the generic S3 functions \code{print()},
\code{summary()}, and \code{mean()}.
\itemize{
\item \code{print()}: prints a single result, by default the first converted rate.
Others can be printed by passing the \code{pos} input. e.g. \code{print(x, pos = 2)}
\item \code{summary()}: prints summary table of all converted rates and metadata, or
those specified by the \code{pos} input. e.g. \code{summary(x, pos = 1:5)}. The summary
can be exported as a separate dataframe by passing \code{export = TRUE}.
\item \code{mean()}: calculates the mean of all converted rates, or those specified by
the \code{pos} input. e.g. \code{mean(x, pos = 1:5)} The mean can be exported as a
separate value by passing \code{export = TRUE}.
}
}
}
\examples{
# Convert a single numeric rate to an absolute rate

# Convert a single numeric rate to a mass-specific rate

# Convert a single numeric rate to an area-specific rate

# Convert a single rate derived via calc_rate to mass-specific

# Convert multiple rates derived via auto_rate to area-specific

}
