% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.R
\name{convert_rate}
\alias{convert_rate}
\title{Convert units of dissolved oxygen.}
\usage{
convert_rate(x, from = NULL, to = NULL, S = 35, t = 25, P = 1.013253)
}
\arguments{
\item{x}{numeric vector or object of class \code{calc.rate}, \code{auto.rate} or \code{adjust.rate}.}

\item{from}{string.}

\item{to}{string.}

\item{S}{numeric. Salinity. Defaults to 35.}

\item{t}{numeric. Temperature. Defaults to 25 (degrees celcius).}

\item{P}{numeric. Pressure. Defaults to 1.013253 (kPa).}
}
\value{
A list.
}
\description{
Convert units of dissolved oxygen.
}
\examples{
x <- calc.rate(sardine.rd)
convert_rate(x, from = "percent", to = "mg L-1")

convert_rate(100, from = "percent", to = "mg L-1", S = 33, t = 18)

convert_rate(sardine.rd[[2]], from = "percent", to = "torr")
}
