% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_val.R
\name{convert_val}
\alias{convert_val}
\title{Convert values to a different unit}
\usage{
convert_val(x, from = NULL, to = NULL)
}
\arguments{
\item{x}{numeric value or vector. Values to be converted to a different unit.}

\item{from}{string. Unit of the original values.}

\item{to}{string. Unit to be converted to. These defaults are applied if left
NULL: volume "L", temperature "C", mass "kg", area "m2", pressure "bar".}
}
\value{
Output is a numeric vector of converted values.
}
\description{
This is a basic function that converts values of temperature, volume, mass,
area, and atmospheric pressure to different units. This can be useful in
\code{\link[=convert_DO]{convert_DO()}}, \code{\link[=convert_rate]{convert_rate()}}, and \code{\link[=convert_rate.ft]{convert_rate.ft()}} in which some
inputs must be in specific units (e.g. atmospheric pressure in bar, area in
m2). See examples.
}
\details{
If the \code{'to'} input is left \code{NULL}, these defaults are applied depending on
the unit type of the \code{from} input: volume "L", temperature "C", mass "kg",
area "m2", pressure "bar".

A fuzzy string matching algorithm is used to accept different unit formatting
styles. For example, \verb{'msq', 'm2', 'M2', 'sqm'} are all parsed as metres
squared of area.

\strong{Units Accepted}

\emph{Temperature:}
\itemize{
\item "C", "K", "F"
}

\emph{Pressure:}
\itemize{
\item "kPa", "hPa", "Pa", "ubar", "mbar", "bar", "atm" (standard atmospheres),
"Torr"
}

\emph{Volume:}
\itemize{
\item "uL", "mL", "L"
}

\emph{Mass:}
\itemize{
\item "ug", "mg", "g", "kg"
}

\emph{Area:}
\itemize{
\item "mm2", "cm2", "m2", "km2"
}
}
\examples{
# Convert volume
convert_val(10, "ml", "L")
convert_val(10:15, "ml", "L")

# Convert temperature
convert_val(-273.15, "K", "C")
convert_val(-40, "F", "C")
convert_val(c(2,4,6,8), "C", "F")

# Convert pressure
convert_val(1, "atm", "bar")
convert_val(1010, "hpa", "bar")
convert_val(735, "torr", "kpa")

# Convert area
convert_val(100, "cm2", "m2")
convert_val(10000, "mm2", "cm2")

# Convert mass
convert_val(200, "g", "kg")
convert_val(10000, "ug", "mg")

# Use directly in a respR function which requires inputs to be
# in a specific unit. For example, in convert_rate() pressure
# must be in 'bar' and respirometer volume in 'L'.
# Here, chamber volume is 200 ml, pressure measured in mbar.
x <- suppressWarnings(inspect(urchins.rd, 1, 2))
rate <- calc_rate(x, from = 20, to = 30)
convert_rate(rate, o2.unit = "ml/l", time.unit = "min",
             output.unit = "mg/h",
             volume = convert_val(200, "ml", "L"),
             S = 35, t = 15,
             P = convert_val(1010, "mbar", "bar"))

# Note, the default 'to' units are set to those respR requires in
# these functions ("L" and "bar" here), so do not necessarily need
# to be specified:
x <- suppressWarnings(inspect(urchins.rd, 1, 2))
rate <- calc_rate(x, from = 20, to = 30)
convert_rate(rate, o2.unit = "ml/l", time.unit = "min",
             output.unit = "mg/h",
             volume = convert_val(200, "ml"),
             S = 35, t = 15,
             P = convert_val(1010, "mbar"))
}
