% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.ft.R
\name{flow_unit_parse}
\alias{flow_unit_parse}
\title{Extracts time and volume units from flowrate unit already parsed by verify_units}
\usage{
flow_unit_parse(unit, which)
}
\arguments{
\item{unit}{flowrate unit input to be parsed}

\item{which}{parse which component of unit? "time" or "vol"}
}
\description{
Extracts time and volume units from flowrate unit already parsed by verify_units
}
\keyword{internal}
