% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowthrough.rd}
\alias{flowthrough.rd}
\title{Flowthrough respirometry data}
\format{A data frame}
\usage{
flowthrough.rd
}
\description{
Dissolved oxygen units: ??
Temperature: ??
Salinity: ??
Atm. Pressure: ??
Time units: seconds
Chamber volume (L): ??
Specimen wet mass (g): ??
}
\details{
Conducted at ??.
}
\author{
Nicholas Carey
}
\keyword{datasets}
