% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowthrough.rd}
\alias{flowthrough.rd}
\title{Flowthrough respirometry data on the chiton, \emph{Mopalia lignosa}}
\format{
A data frame object consisting of 935 rows (approx 16 mins of
data),and 4 columns: time, oxygen inflow and outflow concentrations, and
oxygen delta (the outflow minus inflow concentrations).
}
\usage{
flowthrough.rd
}
\description{
A single experiment on the chiton species \emph{Mopalia lignosa} in a custom-built
flowthrough respirometry system. Conducted at University of British Columbia,
Vancouver, BC, Canada.
}
\details{
\itemize{
\item Dissolved oxygen units: \code{mg/L}
\item Time units: \code{seconds}
\item Flow rate (\code{mL/min}): \code{2.34}
\item Inflow oxygen concentration (calculated assuming 100\% air saturated, \code{mg/L}): \code{8.919}
\item Specimen ash-free dry mass (\code{kg}): \code{0.000070}
\item Temperature (°C): \code{t = 12}
\item Salinity: \code{S = 30}
\item Atm. Pressure (bar): \code{P = 1.013}
}
}
\author{
Nicholas Carey
}
\keyword{datasets}
