% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowthrough.rd}
\alias{flowthrough.rd}
\title{Flowthrough respirometry data on the chiton, \emph{Mopalia lignosa}}
\format{
A data frame object consisting of and 935 rows (approx 16 mins of
data),and 4 columns: time, oxygen inflow and outflow concentrations, and
oxygen delta (the difference between inflow and outflow concentrations).
}
\usage{
flowthrough.rd
}
\description{
A single experiment on the chiton species \emph{Mopalia lignosa} in a custom-built
flowthrough respirometry system. Conducted at University of British Columbia,
Vancouver, BC, Canada.
}
\details{
\itemize{
\item Dissolved oxygen units: mg/L
\item Time units: seconds
\item Flow rate (L/min): 0.00234
\item Inflow oxygen concentration (mean, mg/L): 8.88
\item Specimen ash-free dry mass (kg): 0.000070
}
}
\author{
Nicholas Carey
}
\keyword{datasets}
