% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowthrough_mult.rd}
\alias{flowthrough_mult.rd}
\title{Multi-column flowthrough respirometry data}
\format{
A data frame object consisting of 3740 rows (approx 62 mins of
data),and 13 columns: time (col 1), oxygen outflow concentrations (cols
2,3,4), inflow concentrations (cols 5,6,7, each paired with outflow
columns), delta concentrations (cols 8,9,10, or difference between outflow
and inflow concentrations), inflow concentrations recorded in a shared
header tank (col 11), and outflow and inflow concentrations from a control
or "blank" chamber (cols 12,13).
}
\usage{
flowthrough_mult.rd
}
\description{
A simulated dataset for testing and demonstrating flowthrough respirometry
analyses. Contains one column of numeric time data (s), three columns of
outflow oxygen concentrations (oxy.out.1, etc), three columns of inflow
oxygen concentrations (oxy.in.1, etc.), and three columns of delta oxygen
concentrations (oxy.delta.1, etc., which is simply the numeric difference
between paired columns of out.o2 and in.o2). There is also a column of inflow
oxygen concentrations as recorded from a shared header tank (oxy.header)
supplying all chambers, to use as an alternative to the individual inflow
oxygen recordings. In addition, there are two columns of recordings from an
empty control respirometer, or "blank" experiment (oxy.out.blank,
oxy.in.blank) to determine background respiration.
}
\details{
The third dataset (col 4) has a period of higher rates at around the 2000s
timepoint, where the specimen increases its activity then slowly recovers to
routine respiration levels.
\itemize{
\item Dissolved oxygen units: mg/L
\item Time units: seconds
}
}
\author{
Nicholas Carey
}
\keyword{datasets}
