% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import from sensor output file}
\usage{
import_file(path)
}
\arguments{
\item{path}{string. Path to file.}
}
\value{
a data frame object of time (absolute)
}
\description{
Automatically read and clean data from sensor output. The aim is to work with
most commercial DO sensors available in the market with little input from the
user. This is done by performing simple regular expressions to identify the
file source by extracting the first line of the file and matching unique
strings.
}
\examples{
NULL
}
