% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import from sensor output file}
\usage{
import_file(path, export = FALSE)
}
\arguments{
\item{path}{string. Path to file.}

\item{export}{logical. If TRUE, saves the file in the same directory,
determined by the path parameter above.}
}
\value{
a data frame object of time (absolute)
}
\description{
Automatically read and clean data from sensor output. The aim is to work with
most commercial DO sensors available in the market with little input from the
user. This is done by performing simple regular expressions to identify the
file source by extracting the first line of the file and matching unique
strings. It's a simple procedure for now, but once we have a large database
of files we will optimise the code.
}
\details{
Currently works for: Firesting Logger | Pyro Oxygen Logger (also Firesting) |
PRESENS OXY10 | MiniDOT | Loligo Witrox Logger | Loligo AutoResp (software
output)

While the devices listed above are supported, the import functionality may
not be complete due to limited access to output files. This will improve over
time as users provide feedback. We are releasing this as it is, without any
warranty, so that some people can still benefit from the functionality as it
gets better. Users should still be expected to be able to import data by
themselves since it is a fundamental skill in R.
}
\examples{
NULL
}
