% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.data.R
\name{inspect.data}
\alias{inspect.data}
\title{Title}
\usage{
inspect.data(df, xcol = 1, ycol = 2, highlight = TRUE, plot = TRUE)
}
\arguments{
\item{df}{data frame.}

\item{xcol}{numeric.}

\item{ycol}{numaric.}

\item{highlight}{logical. Defaults to TRUE. Prints location (row #) of errors.}

\item{plot}{logical. Defaults to TRUE.}
}
\description{
Title
}
