% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.ft.R
\name{inspect.ft}
\alias{inspect.ft}
\title{Explore and visualise flowthrough respirometry data and check for errors}
\usage{
inspect.ft(
  x,
  time = NULL,
  out.o2 = NULL,
  in.o2 = NULL,
  in.o2.value = NULL,
  delta.o2 = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} containing columns of \code{time} and \code{out.o2} or \code{delta.o2}
concentrations, and optionally \code{in.o2}.}

\item{time}{integer. Defaults to 1. Specifies the column number of the time
data.}

\item{out.o2}{integer(s). Defaults to \code{NULL}. Specifies the column number(s)
of outflow O2 data.}

\item{in.o2}{integer(s). Defaults to \code{NULL}. Specifies the column number(s)
of inflow O2 data.}

\item{in.o2.value}{numeric value. Defaults to \code{NULL}. If there is no
continuous \code{in.o2} data, this specifies a fixed value of oxygen
concentration for inflowing water in same units as \code{out.o2}, and is used
with \code{out.o2} to calculate a \code{delta.o2}.}

\item{delta.o2}{integer(s). Defaults to all non-time columns if no other
inputs given. Specifies the column number(s) of delta O2 data, for when the
user has already calculated the difference between outflow and inflow
oxygen (should be negative values for oxygen uptake). If this is used,
\code{out.o2} and \code{in.o2} should be NULL.}

\item{plot}{logical. Defaults to TRUE. Plots the data. See Details.}

\item{...}{Allows additional plotting controls to be passed, such as \code{legend = FALSE}, \code{message = FALSE}, \code{rate.rev = FALSE} and \code{pos}.}
}
\description{
\code{inspect.ft} is a data exploration and preparation function that visualises
flowthrough respirometry data, scans it for common issues, and prepares it
for use in later functions in \code{respR}, such as \code{\link[=calc_rate.ft]{calc_rate.ft()}}.
}
\details{
\code{inspect.ft} is intended to be specific to \emph{flowthrough respirometry} data.
In flowthrough respirometry (also known as 'open flow', 'open system' or
'continuous flow' respirometry) rather than calculating a rate from a
changing oxygen concentration recording in a sealed chamber, instead the
difference (i.e. 'oxygen delta') between the inflowing and outflowing oxygen
concentrations of a respirometer receiving water at a constant flow rate is
used to calculate an oxygen consumption or production rate, typically after
it has reached a steady state. Therefore, in general, regions of stable
oxygen delta values (difference between outflow and inflow oxygen) are of
interest. \code{inspect.ft} visualises and prepares the data for use in
\code{\link[=calc_rate.ft]{calc_rate.ft()}}. By specifying data types in this function and saving the
output, they do not need to be specified in later functions.
\subsection{Inputs}{

\code{inspect.ft} requires at least two data inputs; a single column of numeric
\code{time} data, with \emph{either} a column of paired \code{out.o2} concentrations (i.e.
the exhalent or 'downstream' concentrations), \emph{or} a column of already
calculated \code{delta.o2} values, that is the difference between outflow and
inflow concentrations, or the outflow concentration corrected by a background
recording from a 'blank' or empty chamber.

\emph{out.o2 input option}: If an \code{out.o2} column has been specified, in order to
calculate the oxygen delta (and therefore a rate in \code{\link[=calc_rate.ft]{calc_rate.ft()}}) there
must also be an inflow oxygen concentration input (i.e. the inhalent or
'upstream' concentration). This will generally be a column of paired \code{in.o2}
concentrations, in which case the paired values of \code{out.o2} and \code{in.o2} are
used to calculate the oxygen \code{delta.o2}, which is saved in the output, and
used to determine a rate in \code{\link[=calc_rate.ft]{calc_rate.ft()}}. Alternatively, if the inflow
oxygen concentration is a known, generally unvarying concentration (such as
fully air-saturated water from a header tank) this can be entered as a single
value via \code{in.o2.value} and this is used to calculate the \code{delta.o2}.

\emph{delta.o2 input option}: If delta oxygen values have already been calculated,
these can be entered via the \code{delta.o2} input, and these are prepared and
saved for rate calculations in \code{calc_rate.ft}.
}

\subsection{Data processing and validation checks}{

Given an input data frame \code{x}, the function scans the columns specified via
the \code{time}, \code{out.o2}, \code{in.o2} or \code{delta.o2} inputs. If no columns are
specified, by default the functions assumes the first column is \code{time}, and
all others are \code{delta.o2} oxygen data.

The \code{time} column is checked for missing (\code{NA/NaN}) values, that values are
sequential, that there are no duplicate times, and that it is numerically
evenly-spaced. Oxygen columns are simply checked for missing (\code{NA/NaN}) data.
See \strong{Failed Checks} section for what it means for analyses if these checks
result in warnings.
}

\subsection{Plot}{

If \code{plot = TRUE}, entered data is plotted against both time (bottom, blue
axis) and row index (top, red axis), depending on the inputs:
\itemize{
\item a single \code{out.o2} column with either a paired \code{in.o2} column or
\code{in.o2.value}: a two panel plot. The top plot is both outflow (black points)
and inflow (grey points) oxygen. The bottom plot is the oxygen delta between
outflow and inflow oxygen, essentially the oxygen uptake or production rate.
\item a single \code{delta.o2} column: a one panel plot of oxygen delta values.
\item multiple \code{out.o2} or \code{delta.o2} columns: a grid plot of all \code{delta.o2} data
(either as entered or calculated from \code{out.o2} and \code{in.o2}). Specific delta
plots can be examined individually by using the \code{pos} input (e.g. \code{plot(x, pos = 2)}). Y-axes are not equal.
\item unspecified columns: all columns are plotted assuming \code{time} is in column
1, and all others are oxygen \code{delta.o2} data. Y-axes are not equal.
}

In delta plots, that is those plotting \code{delta.o2} values, either directly
entered or calculated, consistent oxygen uptake or production rates will be
represented by flat or level regions.

\emph{\strong{Note:}} Since \code{respR} is primarily used to examine oxygen consumption,
the delta oxygen and rate plots are by default plotted on a reverse y-axis.
In \code{respR} oxygen uptake rates are negative since they represent a negative
slope of oxygen against time. In these plots the axis is reversed so that
higher uptake rates (i.e. more negative) will be higher on these plots. If
you are interested instead in oxygen production rates, which are positive,
the \code{rate.rev = FALSE} argument can be passed in either the \code{inspect.ft}
call, or when using \code{plot()} on the output object. In this case, the delta
and rate values will be plotted numerically, with higher oxygen \emph{production}
rates higher on the plot.

If the legend or labels obscure part of the plot, they can be suppressed via
\code{legend = FALSE} in either the \code{inspect.ft} call, or when using \code{plot()} on
the output object.
}

\subsection{Multiple data columns}{

For a quick overview of larger experiments, multiple columns of \code{out.o2},
\code{in.o2} and \code{delta.o2} can be inspected, but must share the same numeric time
data column specified by the \code{time} input. Note, multiple column inspection
is chiefly intended to be exploratory functionality to provide a quick
overview of larger datasets. While the output will contain all data columns
in \verb{$dataframe} and \verb{$data}, subsequent functions such as \code{\link[=calc_rate.ft]{calc_rate.ft()}}
will use only the first \code{delta.o2} column for calculating rates. Best
practice is to inspect and assign each individual experiment or column pair
as separate \code{inspect.ft} objects. See Examples.

If multiple \code{out.o2} columns are specified, \code{in.o2} can be a single column
(if for example all chambers are supplied from the same header tank), in
which case it is used to calculate an oxygen delta for all \code{out.o2} columns.
A single \code{in.o2.value} in the same units as \code{out.o2} can also be specified.
There can also be multiple \code{in.o2} columns, in which case it is assumed each
\code{out.o2} column is paired with each \code{in.o2} at the same position, and used to
calculate the oxygen \code{delta.o2}. Therefore, \code{out.o2} and \code{in.o2} must have
equal numbers of columns.
}

\subsection{Failed Checks}{

It should be noted the data checks in \code{inspect.ft} are mainly for exploratory
purposes; they help diagnose and flag potential issues with the data. For
instance, long experiments may have had sensor dropouts the user is unaware
of. Others are not really issues at all. For instance, an uneven time warning
can result from using decimalised minutes, which is a completely valid time
metric, but happens to be numerically unevenly spaced. As an additional
check, if uneven time is found, the minimum and maximum intervals in the time
data are in the console output, so a user can see immediately if there are
large gaps in the data.

If some of these checks fail, it should \emph{generally} not hinder analysis of
the data. \code{respR} has been coded to rely on linear regressions on exact data
values, and not make assumptions about data spacing or order. Therefore
issues such as missing or NA/NaN values, duplicate or non-sequential time
values, or uneven time spacing should not cause any erroneous results, as
long as they do not occur over large regions of the data. \code{inspect.ft}
however outputs locations (row numbers) of where these issues occur (located
in the \verb{$locs} element of the output), allowing users to amend them before
analysis. We would recommend that to be completely confident in any results
from analysis of such data, and avoid obscure errors, these issues be
addressed before proceeding.
}

\subsection{Background control or "blank" experiments}{

For experiments in which the specimen data is to be background corrected by a
concurrently-run control experiment, \code{inspect.ft} can be used by specifying
the specimen experiment as \code{out.o2}, and the "blank" as the \code{in.o2} input. In
this way, any variations in oxygen in the specimen data due to background
microbial activity, or for any other reason such as fluctuations in inflow
oxygen, are accounted for in the delta oxygen calculations, and therefore in
the rate calculated in \code{\link[=calc_rate.ft]{calc_rate.ft()}}. See the vignettes on the website
for examples.

If the background recordings are experiments with their own outflow and
inflow recordings, which show a generally consistent oxygen delta due to
microbial activity, this can be saved as a separate \code{inspect.ft} object, a
background rate calculated in \code{\link[=calc_rate.ft]{calc_rate.ft()}}, and this used in
\code{\link[=adjust_rate.ft]{adjust_rate.ft()}} as the \code{by} input to perform background adjustments to
specimen rates.

\strong{Note}: All background calculations should be from experiments done at the
\emph{same flow rate} as the specimen experiments to be corrected.
}

\subsection{Output}{

Output is a \code{list} object of class \code{inspect.ft} containing input parameters
and data, data check summaries, and metadata, which can be passed to
\code{\link[=calc_rate.ft]{calc_rate.ft()}} to determine rates. If there are failed checks or
warnings, the row locations of the potentially problematic data can be found
in \verb{$locs}.
}

\subsection{S3 Generic Functions}{

Saved output objects can be used in the generic S3 functions \code{print()} and
\code{summary()}.
\itemize{
\item \code{print()}: prints a summary of the checks performed on the data. If issues
are found, locations (row numbers) are printed (up to first 20 occurrences).
\item \code{summary()}: simple wrapper for \code{print()} function. See above.
}
}
}
\examples{

# inspect outflow and inflow O2 data
x <- inspect.ft(flowthrough.rd, time = 1, out.o2 = 2,
                in.o2 = 3)
print(x)
plot(x)

# inspect outflow O2 data with inflow as a known value
x <- inspect.ft(flowthrough.rd, time = 1, out.o2 = 2,
                in.o2.value = 8.9)
print(x)
plot(x)

# inspect already calculated delta O2 data
inspect.ft(flowthrough.rd, time = 1, delta.o2 = 4)

# inspect multiple columns for a quick overview
inspect.ft(flowthrough_mult.rd, time = 1, delta.o2 = 8:10)

# Inspect outflow and use a blank control chamber as background
# correction
#
# This experiment has increasing background respiration over time.
# Inspecting outflow O2 with inflow header tank concentrations
# suggests specimen rates (bottom delta.o2 plot) are increasing.
inspect.ft(flowthrough_nlbg.rd, time = 1,
           out.o2 = 2, in.o2 = 4)

# However, inspecting with recordings from a concurrent blank
# control accounts for this and shows specimen rates are level
# when background is taken into account.
inspect.ft(flowthrough_nlbg.rd, time = 1,
           out.o2 = 2, in.o2 = 3)
}
