% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data.R
\name{inspect_data}
\alias{inspect_data}
\title{Check for common errors in respirometry data}
\usage{
inspect_data(df, xcol = 1, ycol = 2, highlight = TRUE, plot = TRUE)
}
\arguments{
\item{df}{data frame. Accepts data frame object of any size.}

\item{xcol}{numeric. Defaults to \code{1}. This is the time data.}

\item{ycol}{numaric. Defaults to \code{2}. This is the dissolved oxygen data.}

\item{highlight}{logical. Defaults to TRUE. Prints location (row #) of errors
detected by the function.}

\item{plot}{logical. Defaults to TRUE. Produces plots for quick visual
diagnostics.}
}
\value{
A list object of class \code{adjust_rate}.
}
\description{
\code{inspect_data()} scans a data frame for specific errors that may affect the
use of functions in \code{respR}. Data checks include: * A test for NA/NaN inputs.
\itemize{
\item A test for numeric data. * A test for sequential time data. * A test for
duplicate time data. * A test for evenly-spaced time data.
}
}
\details{
Once data checks are complete, the function produces a list object which may
be directly loaded into \code{\link[=calc_rate]{calc_rate()}}, \code{\link[=calc_rate.bg]{calc_rate.bg()}}, \code{\link[=auto_rate]{auto_rate()}} and
\code{\link[=pcrit]{pcrit()}} for further analyses.
}
\examples{
inspect_data(sardine.rd)
inspect_data(urchins.rd, 1, 5, highlight = FALSE)

# It is also possible to load the function directly into respR's
# other functions:
calc_rate(inspect_data(sardine.rd, highlight = FALSE, plot = FALSE),
          from = 3000, to = 4000, by = "time")
}
