% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data.R
\name{inspect_data}
\alias{inspect_data}
\title{Check for common errors in respirometry data}
\usage{
inspect_data(
  df,
  time = NULL,
  oxygen = NULL,
  inflow.o2 = NULL,
  outflow.o2 = NULL,
  highlight = TRUE,
  plot = TRUE
)
}
\arguments{
\item{df}{data frame. Accepts data frame object of any size.}

\item{time}{numeric. Defaults to NULL. This is the time data.}

\item{oxygen}{numeric. Defaults to NULL. This is the dissolved oxygen data.}

\item{inflow.o2}{numeric. Defaults to NULL. This is inflow oxygen data. Used
only for flowthrough respirometry data.}

\item{outflow.o2}{numeric. Defaults to NULL. This is outflow oxygen data.
Used only for flowthrough respirometry data.}

\item{highlight}{logical. Defaults to TRUE. Prints location (row #) of errors
detected by the function.}

\item{plot}{logical. Defaults to TRUE. Produces plots for quick visual
diagnostics.}
}
\value{
A list object of class \code{adjust_rate}.
}
\description{
PLEASE NOTE: the \code{inspect_data} function is \strong{deprecated}. It will not be
updated, and will be removed in a future update to \code{respR}. Please use the
\code{\link{inspect}} function instead.
}
\details{
\code{inspect_data()} scans a data frame for specific errors that may affect the
use of functions in \code{respR}.

Data checks include:
\itemize{
\item A test for NA/NaN inputs. * A test for numeric data. * A test for
sequential time data. * A test for duplicate time data. * A test for
evenly-spaced time data.
}

Once data checks are complete, the function produces a list object which may
be directly loaded into \code{\link[=calc_rate]{calc_rate()}}, \code{\link[=calc_rate.bg]{calc_rate.bg()}}, \code{\link[=calc_rate.ft]{calc_rate.ft()}},
and \code{\link[=auto_rate]{auto_rate()}} for further analyses.
}
\examples{
data("sardine.rd")
inspect_data(sardine.rd)

data("urchins.rd")
inspect_data(urchins.rd, 1, 5, highlight = FALSE)
}
