% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intermittent.rd}
\alias{intermittent.rd}
\title{Respirometry data of the sea urchin, \emph{Heliocidaris Erythrogramma}}
\format{
A data frame object consisting of 2 columns (time and dissolved
oxygen) and 4831 rows (approx 80 min of data).
}
\usage{
intermittent.rd
}
\description{
Multiple measurements of oxygen consumption in a single sea urchin,
\emph{Heliocidaris erythrogramma}, obtained using intermittent flow respirometry.
The experiment was conducted at the Sydney Institute of Marine Science in
Sydney, Australia. There are a total of 3 replicates showing declining oxygen,
separated by flushes where new water was added showing increasing oxygen.
Data was collected using a Vernier Optical DO probe (ODO-BTA).
}
\details{
\itemize{
\item Dissolved oxygen units: mg/L - Time units: seconds - Chamber volume (L):
2.379 - Specimen ash-free dry mass (kg): 0.006955
}
}
\author{
Nicholas Carey
}
\keyword{datasets}
