% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{kde_fit}
\alias{kde_fit}
\title{Perform kernel density estimate and fitting}
\usage{
kde_fit(dt, roll, width, by)
}
\description{
This is an internal function and is used by \code{\link[=auto.rate]{auto.rate()}}.
}
\keyword{internal}
