% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{mean.auto_rate}
\alias{mean.auto_rate}
\title{Average auto_rate object rates}
\usage{
\method{mean}{auto_rate}(x, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{auto_rate object}

\item{pos}{integer(s). Which result(s) to average.}

\item{export}{logical. Export averaged values as single value.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Average auto_rate object rates
}
\keyword{internal}
