% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.ft.R
\name{mean.calc_rate.ft}
\alias{mean.calc_rate.ft}
\title{Average calc_rate.ft object rates}
\usage{
\method{mean}{calc_rate.ft}(x, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{calc_rate.ft object}

\item{pos}{integer(s). Which result(s) to average.}

\item{export}{logical. Export averaged values as single value.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Average calc_rate.ft object rates
}
\keyword{internal}
