% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.R
\name{mean.convert_rate}
\alias{mean.convert_rate}
\title{Average convert_rate object rates}
\usage{
\method{mean}{convert_rate}(x, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{convert_rate object}

\item{pos}{integer(s). Which result(s) to average.}

\item{export}{logical. Export averaged values as single value.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Average convert_rate object rates
}
\keyword{internal}
