% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy_crit.R
\name{mean.oxy_crit}
\alias{mean.oxy_crit}
\title{Average oxy_crit object rates}
\usage{
\method{mean}{oxy_crit}(x, ...)
}
\arguments{
\item{x}{oxy_crit object}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Average oxy_crit object rates
}
\keyword{internal}
