% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrit.R
\name{pcrit}
\alias{pcrit}
\title{Calculate critical oxygen tension for respirometry}
\usage{
pcrit(df, span = 0.05, datmr = FALSE, plot = T)
}
\arguments{
\item{df}{data frame object}

\item{span}{numeric}

\item{MR}{logical}
}
\description{
Uses stepwise linear regression to calculate two best-fit lines in a rate data frame, by minimising the total residual sum of squares between both regressions. Method is based on Yeager and Ultsch (1989) to calculate critical oxygen tension. Calling \code{pcrit} on a data frame \code{df} of dissolved oxygen (DO) by time will prompt the function to first perform a rolling regression of width \code{floor(0.05 * nrow(df))} to obtain the necessary metabolic rate (MR) data. If the user already has MR data, calling the function with the argument \code{datmr = TRUE} will perform the stepwise regressions immediately without performing rolling regressions.
}
\examples{
# load example data:
data(fishpcrit)
# run pcrit:
pcfish <- pcrit(fishpcrit)
print(pcfish)              # print the rank 1 result
print(pcfish, rank = 1000) # print the rank 1000 result (just to compare)

plot(pcfish)               # plot the rank 1 result
plot(pcfish, rank = 1000)  # plot the rank 1000 results (just to compare)

summary(pcfish)            # prints the top 6 results
summary(pcfish, n = 20)    # prints the top 20 results (just to compare)


}
