% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrit.R
\name{pcrit}
\alias{pcrit}
\title{Calculate pcrit}
\usage{
pcrit(df, span = 0.05, datmr = FALSE, plot = T)
}
\arguments{
\item{df}{data frame object}

\item{span}{numeric}

\item{MR}{logical}
}
\description{
I need a description for this...
}
\examples{
NULL
}
