% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrit.R
\name{pcrit}
\alias{pcrit}
\title{Calculate Pcrit}
\usage{
pcrit(df, span = 0.05, MR = FALSE)
}
\arguments{
\item{df}{Data frame.}

\item{span}{Numeric.}

\item{MR}{Logical.}
}
\value{
A value of Pcrit.
}
\description{
Calculate Pcrit
}
