% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{plot.auto_rate}
\alias{plot.auto_rate}
\title{Plot auto_rate objects}
\usage{
\method{plot}{auto_rate}(
  x,
  pos = 1,
  panel = FALSE,
  quiet = FALSE,
  legend = TRUE,
  rate.rev = TRUE,
  ...
)
}
\arguments{
\item{x}{auto_rate object}

\item{pos}{integer. Which result to plot.}

\item{panel}{integer. Which panel to plot individually.}

\item{quiet}{logical. Suppress console output.}

\item{legend}{logical. Suppress labels and legends.}

\item{rate.rev}{logical. Control direction of y-axis in rolling rate plot.}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot auto_rate objects
}
\keyword{internal}
