% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.bg.R
\name{plot.calc_rate.bg}
\alias{plot.calc_rate.bg}
\title{Plot calc_rate.bg objects}
\usage{
\method{plot}{calc_rate.bg}(x, pos = NULL, quiet = FALSE, legend = TRUE, ...)
}
\arguments{
\item{x}{calc_rate.bg object}

\item{pos}{integer. Which result to plot.}

\item{quiet}{logical. Suppress console output.}

\item{legend}{logical. Suppress labels and legends.}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot calc_rate.bg objects
}
\keyword{internal}
