% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.R
\name{plot.calc_rate}
\alias{plot.calc_rate}
\title{Plot calc_rate objects}
\usage{
\method{plot}{calc_rate}(x, pos = 1, quiet = FALSE, panel = NULL, legend = TRUE, ...)
}
\arguments{
\item{x}{calc_rate.bg object}

\item{pos}{integer. Which result to plot.}

\item{quiet}{logical. Suppress console output.}

\item{panel}{integer. Which panel to plot individually.}

\item{legend}{logical. Suppress labels and legends.}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot calc_rate objects
}
