% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.ft.R
\name{plot.convert_rate.ft}
\alias{plot.convert_rate.ft}
\title{Plot convert_rate.ft objects}
\usage{
\method{plot}{convert_rate.ft}(x, ...)
}
\arguments{
\item{x}{convert_rate.ft object}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot convert_rate.ft objects
}
