% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.R
\name{plot.convert_rate}
\alias{plot.convert_rate}
\title{Plot convert_rate objects}
\usage{
\method{plot}{convert_rate}(x, ...)
}
\value{
A plot. No returned value.
}
\description{
Plot convert_rate objects
}
