% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.ft.R
\name{plot.inspect.ft}
\alias{plot.inspect.ft}
\title{Plot inspect.ft objects}
\usage{
\method{plot}{inspect.ft}(
  x,
  width = NULL,
  pos = NULL,
  quiet = FALSE,
  legend = TRUE,
  rate.rev = TRUE,
  ...
)
}
\arguments{
\item{x}{inspect.ft object}

\item{width}{numeric. Smoothing factor (rolling mean) for delta oxygen values
as proportion of total data length (0 to 1)}

\item{pos}{integer. Which result to plot.}

\item{quiet}{logical. Suppress console output.}

\item{legend}{logical. Suppress labels and legends.}

\item{rate.rev}{logical. Control direction of y-axis in delta oxygen plot.}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot inspect.ft objects
}
\keyword{internal}
