% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{plot.inspect}
\alias{plot.inspect}
\title{Plot inspect objects}
\usage{
\method{plot}{inspect}(
  x,
  width = NULL,
  pos = NULL,
  quiet = FALSE,
  legend = TRUE,
  rate.rev = TRUE,
  ...
)
}
\arguments{
\item{x}{inspect object}

\item{width}{numeric. Width of rolling regression to determine rates in
rolling rate plot as proportion of total data length (0 to 1)}

\item{pos}{integer. Which result to plot.}

\item{quiet}{logical. Suppress console output.}

\item{legend}{logical. Suppress labels and legends.}

\item{rate.rev}{logical. Control direction of y-axis in rolling rate plot.}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot inspect objects
}
