% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy_crit.R
\name{plot.oxy_crit}
\alias{plot.oxy_crit}
\title{Plot oxy_crit objects}
\usage{
\method{plot}{oxy_crit}(x, legend = TRUE, quiet = FALSE, panel = NULL, rate.rev = TRUE, ...)
}
\arguments{
\item{x}{oxy_crit object}

\item{legend}{logical. Suppress labels and legends.}

\item{quiet}{logical. Suppress console output.}

\item{panel}{integer. Which panel to plot individually.}

\item{rate.rev}{logical. Control direction of y-axis in rolling rate plot.}

\item{...}{Pass additional plotting parameters}
}
\value{
A plot. No returned value.
}
\description{
Plot oxy_crit objects
}
\keyword{internal}
