% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.R
\name{print.adjust_rate}
\alias{print.adjust_rate}
\title{Print adjust_rate objects}
\usage{
\method{print}{adjust_rate}(x, pos = 1, ...)
}
\value{
Print to console. No returned value.
}
\description{
Print adjust_rate objects
}
