% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.ft.R
\name{print.convert_rate.ft}
\alias{print.convert_rate.ft}
\title{Print convert_rate.ft objects}
\usage{
\method{print}{convert_rate.ft}(x, pos = 1, ...)
}
\arguments{
\item{x}{convert_rate.ft object}

\item{pos}{integer. Which result to print.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Print convert_rate.ft objects
}
\keyword{internal}
