% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respR.R
\docType{package}
\name{respR-package}
\alias{respR}
\alias{respR-package}
\title{respR: Analyse, convert, and automate respirometry-related data.}
\description{
respR is an R package that consists of a collection of functions that provide
a structured workflow for the analysis of respirometry-related data. Focus is
placed on ease-of-use.
}
\details{
Works with closed, open, intermittent, and interval measurements. Functions
for background corrections and unit conversions are included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/januarharianto/respr}
}

}
\author{
Januar Harianto & Nicholas Carey
}
