% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sardine.rd}
\alias{sardine.rd}
\title{Respirometry data of the sardine, \emph{Sardinops sagax}}
\format{A data frame}
\usage{
sardine.rd
}
\description{
Dissolved oxygen units: %
Temperature: ??
Salinity: 35
Atm. Pressure: ??
Time units: seconds
Chamber volume (L): ??
Specimen wet mass (g): ??
}
\details{
Conducted at Hopkins Marine Station, Stanford University.
}
\author{
Nicholas Carey
}
\keyword{datasets}
