% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sardine.rd}
\alias{sardine.rd}
\title{Respirometry data of the sardine, \emph{Species}}
\format{A data frame}
\usage{
sardine.rd
}
\description{
Respirometry data of the sardine, \emph{Species}
}
\author{
Nicholas Carey
}
\keyword{datasets}
