% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squid.rd}
\alias{squid.rd}
\title{Measurement of P_crit in the squid, \emph{Doryteuthis opalescens}}
\format{A data frame}
\source{
\url{http://www.diamondse.info/}
}
\usage{
squid.rd
}
\description{
Measurement of P_crit in the squid, \emph{Doryteuthis opalescens}
}
\author{
Nicholas Carey
}
\keyword{datasets}
