% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squid.rd}
\alias{squid.rd}
\title{Respirometry data of the squid, \emph{Doryteuthis opalescens}, with P_crit}
\format{A data frame}
\usage{
squid.rd
}
\description{
A single experiment on the squid species \emph{Doryteuthis opalescens} in a
Loligo Systems swim tunnel and Witrox oxygen probe system. Experiment
conducted at Hopkins Marine Station, Stanford University, Pacific Grove,
California.
}
\details{
Dissolved oxygen units: mg/L
Time units: seconds
Chamber volume (L): 12.3
Specimen wet mass (g): 21.41

Data kindly supplied by Ben Burford, Hopkins Marine Station,
Stanford University. \href{mailto:bburford@stanford.edu}{bburford@stanford.edu}
}
\author{
Ben Burford
}
\keyword{datasets}
