% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squid.rd}
\alias{squid.rd}
\title{Respirometry data of the squid, \emph{Doryteuthis opalescens}, with P_crit}
\format{A data frame}
\usage{
squid.rd
}
\description{
Respirometry data of the squid, \emph{Doryteuthis opalescens}, with P_crit
}
\author{
Nicholas Carey
}
\keyword{datasets}
