% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squid.rd}
\alias{squid.rd}
\title{Respirometry data of the squid, \emph{Doryteuthis opalescens}, with P_crit}
\format{A data frame object consisting of 2 columns (time and dissolved
oxygen) and 34120 rows (approx 9.5h of data).}
\usage{
squid.rd
}
\description{
A single experiment on the squid species \emph{Doryteuthis opalescens} in a
Loligo Systems swim tunnel and Witrox oxygen probe system. Experiment
conducted at Hopkins Marine Station, Stanford University, Pacific Grove,
California.
}
\details{
\itemize{
\item Dissolved oxygen units: mg/L
\item Time units: seconds
\item Chamber volume (L): 12.3
\item Specimen wet mass (kg): 0.02141
}

Data kindly supplied by Ben Burford, Hopkins Marine Station,
Stanford University. \href{mailto:bburford@stanford.edu}{bburford@stanford.edu}
}
\author{
Ben Burford
}
\keyword{datasets}
