% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Subsample a data frame object}
\usage{
subsample(x, n = NULL, length.out = NULL, random_start = FALSE, plot = TRUE)
}
\arguments{
\item{x}{data frame or vector. The data to subsample.}

\item{n}{numeric. Subsample every \code{n} rows.}

\item{length.out}{numeric. Subsample to a specific length or number of rows.}

\item{random_start}{logical. Defaults to FALSE. If TRUE, randomises the start
row of the data (applies to \code{n} input only).}

\item{plot}{logical. Defaults to TRUE. Plots the data.}
}
\description{
A simple function that subsamples a data frame of time~oxygen data in order
to "thin" large datasets. It also works on vectors of any data type. Two
subsampling methods are provided. Using the \code{n} argument selects every n
elements or rows. The \code{length.out} argument uniformly subsamples the input to
the desired length.
}
\details{
\subsection{Output}{

Returns a data frame or vector object depending on input.
}
}
\examples{
# Subsample by every 10th row:
subsample(squid.rd, n = 10)

# Subsample to 1000 rows:
subsample(squid.rd, length.out = 1000)

# Subsample with random starting position:
subsample(sardine.rd, n = 3, random_start = TRUE)
}
