% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Subsample a data frame object}
\usage{
subsample(df, n = 5, random_start = F, plot = T)
}
\arguments{
\item{df}{data frame.}

\item{n}{numeric.}

\item{random_start}{logical.}
}
\value{
A data frame object.
}
\description{
A simple function that selects every "\code{n}" rows in a data frame to create a
new data frame object. Effectifely "thins" large datasets.
}
\examples{
# Subsample every 10 rows:
subsample(squid.rd, n = 10)

# Subsample with random first value:
subsample(sardine.rd, 3, T)
}
\author{
Januar Harianto & Nicholas Carey
}
