% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Subsample a data frame object}
\usage{
subsample(df, n = 5, random_start = FALSE, plot = TRUE)
}
\arguments{
\item{df}{data frame. The data to process.}

\item{n}{numeric. This will prompt the function to extract every "\code{n}" rows.}

\item{random_start}{logical. Defaults to FALSE. If TRUE, randomises the start
row of the data.}

\item{plot}{logical. Defaults to TRUE. Plots the data automatically.}
}
\value{
A data frame object.
}
\description{
A simple function that selects every "\code{n}" rows in a data frame to create a
new data frame object. Effectively "thins" large datasets.
}
\examples{
# Subsample every 10 rows:
subsample(squid.rd, n = 10)

# Subsample with random first value:
subsample(sardine.rd, 3, TRUE)
}
