% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{Truncate a data frame to create a subset of the data}
\usage{
subset_data(x, from, to, by)
}
\arguments{
\item{x}{data frame. The data to subset.}

\item{from}{numeric. Defines the upper bound(s) of the data frame to subset.
Subsetting is based on the argument: \code{by}.}

\item{to}{numeric. Defines the lower bound(s) of the data frame to subset.
Subsetting is based on the argument: \code{by}.}

\item{by}{string. "time", "row", "o2" or "proportion".}
}
\value{
A \code{data.table} object.
}
\description{
This is an internal function. This function extracts a subset data frame
based on a given set of rules.
}
