% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rate.ft.R
\name{summary.adjust_rate.ft}
\alias{summary.adjust_rate.ft}
\title{Summarise adjust_rate.ft objects}
\usage{
\method{summary}{adjust_rate.ft}(object, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Summarise adjust_rate.ft objects
}
