% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{summary.auto_rate}
\alias{summary.auto_rate}
\title{Summarise auto_rate objects}
\usage{
\method{summary}{auto_rate}(object, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Summarise auto_rate objects
}
