% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.bg.R
\name{summary.calc_rate.bg}
\alias{summary.calc_rate.bg}
\title{Summarise calc_rate.bg objects}
\usage{
\method{summary}{calc_rate.bg}(object, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{object}{calc_rate.bg object}

\item{pos}{integer(s). Which summary row(s) to print.}

\item{export}{logical. Export summary table as data frame.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise calc_rate.bg objects
}
