% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rate.ft.R
\name{summary.calc_rate.ft}
\alias{summary.calc_rate.ft}
\title{Summarise calc_rate.ft objects}
\usage{
\method{summary}{calc_rate.ft}(object, pos = NULL, export = FALSE, ...)
}
\value{
Print to console. No returned value.
}
\description{
Summarise calc_rate.ft objects
}
