% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.ft.R
\name{summary.convert_rate.ft}
\alias{summary.convert_rate.ft}
\title{Summarise convert_rate.ft objects}
\usage{
\method{summary}{convert_rate.ft}(object, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{object}{convert_rate.ft object}

\item{pos}{integer(s). Which summary row(s) to print.}

\item{export}{logical. Export summary table as data frame.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise convert_rate.ft objects
}
\keyword{internal}
