% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rate.R
\name{summary.convert_rate}
\alias{summary.convert_rate}
\title{Summarise convert_rate objects}
\usage{
\method{summary}{convert_rate}(object, pos = NULL, export = FALSE, ...)
}
\arguments{
\item{object}{convert_rate object}

\item{pos}{integer(s). Which summary row(s) to print.}

\item{export}{logical. Export summary table as data frame.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise convert_rate objects
}
