% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy_crit.R
\name{summary.oxy_crit}
\alias{summary.oxy_crit}
\title{Summarise oxy_crit objects}
\usage{
\method{summary}{oxy_crit}(object, export = FALSE, ...)
}
\arguments{
\item{object}{oxy_crit object}

\item{export}{logical. Export result as data frame.}

\item{...}{Pass additional inputs}
}
\value{
Print to console. No returned value.
}
\description{
Summarise oxy_crit objects
}
\keyword{internal}
