% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rate.R
\name{time_roll}
\alias{time_roll}
\title{Perform time-width rolling regression}
\usage{
time_roll(dt, width)
}
\description{
<<<<<<< HEAD
Perform time-width rolling regression
=======
This is an internal function. Used by \code{\link[=auto_rate]{auto_rate()}}.
>>>>>>> develop
}
\keyword{internal}
