% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_args.R
\name{unit_args}
\alias{unit_args}
\title{Print examples of unit arguments for use in \code{\link[=convert_DO]{convert_DO()}} and
\code{\link[=convert_rate]{convert_rate()}}}
\usage{
unit_args()
}
\value{
A list.
}
\description{
This is a simple function, requiring no argument. It shows available units
that can be used in the functions \code{\link[=convert_DO]{convert_DO()}} and \code{\link[=convert_rate]{convert_rate()}}. Some
units also require temperature (\code{t}), salinity (\code{S}), and atmospheric
pressure (\code{P}) to be specified.
}
\details{
\strong{Available units for \code{\link[=convert_DO]{convert_DO()}}:}

Do \emph{NOT} require t, S and P:

"mg/L", "ug/L", "mmol/L", "umol/L"

Require t, S and P:

"mL/L", "mg/kg", "ug/kg", "mmol/kg",
"umol/kg", "mL/kg", "\\%" (N.B. = \\% Air Saturation), "Torr", "hPa", "kPa",
"mmHg", "inHg"

\strong{Available units for \code{\link[=convert_rate]{convert_rate()}}:}

\strong{O2}

Do \emph{NOT} require t, S and P:

"mg/L", "ug/L", "mmol/L", "umol/L"

Require t, S and P:

"mL/L", "mg/kg", "ug/kg", "mmol/kg",
"umol/kg", "mL/kg", "\\%" (N.B. = \\% Air Saturation), "Torr", "hPa", "kPa",
"mmHg", "inHg"

\strong{Time}

"s", "m", "h"

\strong{Output mass}

"ug", "mg", "g", "kg"
}
\examples{
unit_args()

}
