% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_args.R
\name{unit_args}
\alias{unit_args}
\title{Print examples of unit arguments for use in \code{convert_DO}, \code{convert_rate}, and
\code{convert_rate.ft}}
\usage{
unit_args()
}
\description{
This is a simple function, with no arguments. It prints to the console the
units that can be used in the functions \code{\link[=convert_DO]{convert_DO()}}, \code{\link[=convert_rate]{convert_rate()}}, and
\code{\link[=convert_rate.ft]{convert_rate.ft()}}. Some oxygen unit conversions also require temperature
(\code{t}), salinity (\code{S}), and atmospheric pressure (\code{P}) to be specified.
}
\details{
Note the difference between percent air saturation (\verb{\%Air}), where air
saturated water is ~100\%, and percent oxygen saturation (\verb{\%O2}), where air
saturated water is ~20.946\% \emph{oxygen} saturated. In other words, \verb{\%O2 = \%Air * 0.20946}.

\strong{\code{\link[=convert_DO]{convert_DO()}}}

\strong{Oxygen concentration or pressure units for \code{from} and \code{to} arguments:}

Oxygen concentration units should use SI units (\code{L} or \code{kg}) for the
denominator.

Do \emph{NOT} require \code{t}, \code{S} and \code{P} for conversions:
\itemize{
\item \verb{"mg/L", "ug/L", "mmol/L", "umol/L"}
}

Require \code{t}, \code{S} and \code{P} for conversions:
\itemize{
\item \verb{"mL/L", "mg/kg", "ug/kg", "mmol/kg", "umol/kg", "mL/kg", "\%Air"} (i.e. \%
Air Saturation), \code{"\%O2"} (i.e. \% Oxygen Saturation), \verb{"Torr", "hPa", "kPa",}
\verb{"mmHg", "inHg"}
}

\strong{\code{\link[=convert_rate]{convert_rate()}} and \code{\link[=convert_rate.ft]{convert_rate.ft()}}}

\strong{Oxygen concentration or pressure units for \code{oxy.unit} argument:}

Oxygen concentration units should use SI units (\code{L} or \code{kg}) for the
denominator.

Do \emph{NOT} require \code{t}, \code{S} and \code{P} for conversions:
\itemize{
\item \verb{"mg/L", "ug/L", "mmol/L", "umol/L"}
}

Require \code{t}, \code{S} and \code{P} for conversions:
\itemize{
\item \verb{"mL/L", "mg/kg", "ug/kg", "mmol/kg", "umol/kg", "mL/kg",} \code{"\%Air"} (i.e. \%
Air Saturation), \code{"\%O2"} (i.e. \% Oxygen Saturation), \verb{"Torr", "hPa", "kPa",}
\verb{"mmHg", "inHg"}
}

\strong{Time units for \code{time.unit} or as part of \code{flowrate.unit} argument:}
\itemize{
\item \verb{"sec", "min", "hour", "day"}
}

\strong{Volume units for use as part of \code{flowrate.unit} argument (\code{convert_rate.ft}
only):}

For example, in \code{'ml/min'}, \code{'L/s'}, etc.
\itemize{
\item \verb{"uL", "mL", "L"}
}

\strong{Combining units for \code{output.unit} argument:}

Must be in correct order, with no special characters other than the
separator:
\itemize{
\item Absolute rates: \code{O2/Time} e.g. \code{"mg/s"}, \code{"umol/min"}, \code{"mL/h"}
\item Mass-specific rates: \code{O2/Time/Mass} e.g. \code{"mg/s/ug"}, \code{"umol/min/g"},
\code{"mL/h/kg"}
\item Area-specific rates: \code{O2/Time/Area} e.g. \code{"mg/s/mm2"}, \code{"umol/min/cm2"},
\code{"mL/h/m2"}
}

\strong{Oxygen amount units for use in \code{output.unit} argument:}
\itemize{
\item \verb{"ug", "mg", "umol", "mmol", "mol", "mL"}
}

\strong{Time units for use in \code{output.unit} argument:}
\itemize{
\item \verb{"sec", "min", "hour", "day"}
}

\strong{Mass units for use in \code{output.unit} argument in mass-specific rates:}
\itemize{
\item \verb{"ug", "mg", "g", "kg"}
}

\strong{Area units for use in \code{output.unit} argument in area-specific rates:}
\itemize{
\item \verb{"mm2", "cm2", "m2", "km2"}
}
}
\examples{
unit_args()

}
