% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urchins.rd}
\alias{urchins.rd}
\title{Multi-column respirometry data of the sea urchin, \emph{Heliocidaris
Erythrogramma}, including background respiration}
\format{
A data frame object consisting of one column of time, 16 columns of
urchin oxygen consumption (\code{a} to \code{p}) and 2 columns of background oxygen
consumption (\code{b1} & \code{b2}). There are 271 rows of data spanning 45 minutes.
}
\usage{
urchins.rd
}
\description{
Oxygen consumption data of 16 individual \emph{Heliocidaris erythrogramma}
specimens. In addition, there are two columns of background respiration.
}
\details{
\itemize{
\item Dissolved oxygen units: \code{mg/L}
\item Time units: \code{minutes}
\item Volume (L): \code{1.09}
\item Temperature (°C): \code{20}
\item Salinity: \code{30}
\item Atm. Pressure (bar): \code{1.01}
}
}
\author{
Januar Harianto
}
\keyword{datasets}
